% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_validate.R
\name{add_attr}
\alias{add_attr}
\title{Add specific attributes to a dataframe}
\usage{
add_attr(data, pathogen, adm0)
}
\arguments{
\item{data}{A dataframe to which attributes will be added.}

\item{pathogen}{Name of pathogen passed to an attribute of the same name.
This will be retrieved from \code{assays_df} based on \code{test_id} provided by user.}

\item{adm0}{Unique code of the country passed to an attribute of the same
name. Code is based on \code{serotrackr::regions} object.}
}
\value{
A dataframe with additional attributes.
}
\description{
Add specific attributes to a dataframe
}
\note{
An \code{id} attribute is also added using \code{generate_obj_id()}.
}
\examples{
df <- add_attr(sample_raw_data, pathogen = "SARS-CoV-2",
               adm0 = regions$adm0$Canada)
attributes(df)
}
