% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{all_stds_MFI_csv}
\alias{all_stds_MFI_csv}
\title{Example Serological Dataset}
\format{
A data frame with 3860 rows and 4 variables:
\describe{
  \item{Sample}{Unique identifier for the sample}
  \item{Location}{Bead coupling location for standards}
  \item{MFI}{Median Fluorescent Intensity}
  \item{Antigen}{Antigen name}
}
}
\source{
Mueller Lab, WEHI
}
\description{
MFI values from the WEHI standard curves using both Papua New Guinean
and Ethiopian IgG coupled beads for standard curves.
}
\details{
This file is stored in inst/extdata
}
