% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assays.R
\docType{data}
\name{assays}
\alias{assays}
\title{Assays list}
\format{
\code{assays}
A nested named list
}
\source{
Based on the list of assays gathered by SeroTracker for SARS-CoV-2
}
\usage{
assays
}
\description{
A list of assays for different pathogens. This list can be used in the map_cols() function to select assays.
}
\examples{
assays$`SARS-CoV-2`$`EUROIMMUN - IgG - Anti-SARS-CoV-2 ELISA IgG`

}
\keyword{datasets}
