% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_validate.R
\name{assert}
\alias{assert}
\title{Assert a logical expression}
\usage{
assert(expr, err_msg)
}
\arguments{
\item{expr}{A logical expression to check}

\item{err_msg}{Error message for when the \code{expr} is FALSE}
}
\value{
If \code{expr} is FALSE, it stops the execution and returns the \code{err_msg}.
}
\description{
Assert a logical expression
}
\examples{
f <- function(data) {
  assert("data.frame" \%in\% class(data),
         "Data must be of class 'data.frame'.")
}
f(sample_raw_data)

}
