% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyPv.r
\name{classifyPv}
\alias{classifyPv}
\title{Random Forest Classification for Pv only in Pk/Pf/Pv analysis}
\usage{
classifyPv(mfi_to_rau_output, algorithm_type, sens_spec, counts_QC_output)
}
\arguments{
\item{mfi_to_rau_output}{Output from `MFItoRAU_PNG()` or `MFItoRAU_ETH()`
(reactive).}

\item{algorithm_type}{User-selected algorithm choice:
- "antibody_model" (PvSeroTaT model; default), or
- "antibody_model_excLF016" (PvSeroTaT excluding LF016).}

\item{sens_spec}{User-selected Sensitivity/Specificity threshold:
- "maximised" (default),
- "85\% sensitivity",
- "90\% sensitivity",
- "95\% sensitivity",
- "85\% specificity",
- "90\% specificity".
- "95\% specificity".}

\item{counts_QC_output}{Output from `getCountsQC()` (reactive).}
}
\value{
- Data frame with exposure status for every sample.
- Summary table with positive/negative results for each threshold.
}
\description{
This function classifies unknown samples as recently exposed or not
(Note: MFItoRAU_PNG() or MFItoRAU_ETH() needs to be run first to convert to
RAU). This is a slightly modified function for ONLY use in Pk/Pf/Pv analysis.
The only difference is that ETHtoPNGloglog_Dilution is used to identify the
correct columns to then classify.
}
\author{
Lauren Smith, Dionne Argyropoulos
}
