% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean the data}
\usage{
clean(data)
}
\arguments{
\item{data}{a data.frame}
}
\value{
A cleaned data.frame
}
\description{
Cleans and validates the main columns used during submission to the serotracker platform
}
\examples{
library(dplyr)
library(stringr)
df_orig <- data.frame(dataset_id = 1,
                      id = 1:3,
                      age = c("0 - 10", "20-30", "30,40"),
                      sex = c("f", "Male", "binary"),
                      collection_start_date = "2020-03-01",
                      collection_end_date = "2020-04-01",
                      result = c("'2.4'", "3.5", "9"),
                      result_cat = rep("negative", times = 3))

mapped_cols <- map_cols(
  df_orig,
  dataset_id = dataset_id,
  id = id,
  age_group = age,
  sex = sex,
  country = "Earth!",
  collection_start_date = collection_start_date,
  collection_end_date = collection_end_date,
  test_id = assays$`SARS-CoV-2`$`EUROIMMUN - IgG - Anti-SARS-CoV-2 ELISA IgG`,
  result = result,
  result_cat = result_cat,
  include_others = FALSE
)

clean(mapped_cols)
}
