% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeroData.R
\name{.clean_luminex}
\alias{.clean_luminex}
\title{Helper function to process luminex (Magpix/Intelliflex) data}
\usage{
.clean_luminex(df, row1, row2)
}
\arguments{
\item{file}{String with the raw data path (reactive).}
}
\value{
Cleaned data fame
}
\description{
Helper function to process luminex (Magpix/Intelliflex) data
}
\examples{
\donttest{

your_raw_data <- system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR")
df            <- .read_luminex_file(your_raw_data)
cfg           <- .magpix_version_config("4.2")

row1         <- which(df$xPONENT == "Median")
row2         <- which(df$xPONENT == "Net MFI")

results      <- .clean_luminex(df, row1, row2)

}
}
\author{
Dionne Argyropoulos
}
