% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeroData.R
\name{.extract_luminex_sections}
\alias{.extract_luminex_sections}
\title{Helper function to process luminex sections}
\usage{
.extract_luminex_sections(df, cfg, plt)
}
\arguments{
\item{df}{String with the raw data path (reactive).}

\item{cfg}{Magpix version output of .magpix_version_config().}

\item{plt}{Platform (magpix, intelliflex)}
}
\value{
List of data_raw, results, counts, blanks, stds, run
}
\description{
Helper function to process luminex sections
}
\examples{
your_raw_data <- system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR")
df            <- .read_luminex_file(your_raw_data)
cfg           <- .magpix_version_config("4.2")
section       <- .extract_luminex_sections(df, cfg, "magpix")
}
\author{
Dionne Argyropoulos
}
