% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFItoRAU.R
\name{.merge_mfitorau}
\alias{.merge_mfitorau}
\title{Helper function to add Sample IDs to output.}
\usage{
.merge_mfitorau(df, layout, plate_level)
}
\arguments{
\item{df}{Data frame following 5-parameter logistic function applied.}

\item{layout}{Output from `readPlateLayout()`.}

\item{plate_level}{Specific plate of interest.}
}
\value{
Processed data frame with correct Sample IDs.
}
\description{
Helper function to add Sample IDs to output.
}
\author{
Dionne Argyropoulos
}
