% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFItoRAU.R
\name{.process_antigen_loglog}
\alias{.process_antigen_loglog}
\title{Helper function to fit a 5-parameter logistic standard curve to dilutions}
\usage{
.process_antigen_loglog(
  subset_data,
  antigen,
  dilution,
  s1_concentration,
  s_final_concentration,
  unknown_letters = c("U", "X")
)
}
\arguments{
\item{subset_data}{Data for one plate.}

\item{antigen}{Data for one antigen.}

\item{dilution}{Set of five or ten.}

\item{s1_concentration}{Concentration of highest dilution.}

\item{s_final_concentration}{Concentration lowest dilution.}

\item{unknown_letters}{Bioplex, Magpix or Intelliflex known unknown letters (Default = U and X).}
}
\value{
A list of the model results data frame and model.
}
\description{
Helper function to fit a 5-parameter logistic standard curve to dilutions
}
\author{
Connie Li Wai Suen, Dionne Argyropoulos
}
