% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeroData.R
\name{.read_luminex_file}
\alias{.read_luminex_file}
\title{Helper function to read raw luminex files}
\usage{
.read_luminex_file(file)
}
\arguments{
\item{file}{String with the raw data path (reactive).}
}
\value{
raw data frame
}
\description{
Helper function to read raw luminex files
}
\examples{
your_raw_data <- system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR")
df <- .read_luminex_file(your_raw_data)
}
\author{
Dionne Argyropoulos
}
