% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFItoRAU.R
\name{.setup_mfitorau_inputs}
\alias{.setup_mfitorau_inputs}
\title{Helper function to set up MFI to RAU function}
\usage{
.setup_mfitorau_inputs(df, plate_list, std_point)
}
\arguments{
\item{df}{Output from `readSeroData()`.}

\item{plate_list}{Output from `readPlateLayout()`.}

\item{std_point}{Standard Point Curve: 5 = 5-point curve, 10 = 10-point curve, "PvLDH" for LDH specific curve. Default = 10. Value is an integer.}
}
\value{
A list of processed sero_data, processed plate layout, antigen names, and parameters for standard curve.
}
\description{
Helper function to set up MFI to RAU function
}
\examples{
\donttest{

# Step 0: Load example raw data
your_raw_data <- c(
  system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate2.csv", package = "SeroTrackR")
)
your_plate_layout <- system.file(
  "extdata",
  "example_platelayout_1.xlsx",
  package = "SeroTrackR"
)

# Step 1: Read serology data and plate layout
sero_data  <- readSeroData(your_raw_data,"magpix")
plate_list <- readPlateLayout(your_plate_layout, sero_data)

# Step 2: Setup MFI to RAU
setup <- .setup_mfitorau_inputs(
  df = sero_data$results,
  plate_list = plate_list,
  std_point = 10
)

}

}
\author{
Dionne Argyropoulos
}
