% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{example_BioPlex_plate1_csv}
\alias{example_BioPlex_plate1_csv}
\title{Example Serological Dataset: Bioplex Plate 1}
\format{
A data frame with 103 rows and 15 columns.
}
\source{
Randomised data
}
\description{
A dataset containing raw MFI values and metadata from a sample plate run (Bioplex).
}
\details{
This file is stored in inst/extdata
}
