% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{example_BioPlex_plate3_xlsx}
\alias{example_BioPlex_plate3_xlsx}
\title{Example Serological Dataset: Bioplex Plate 3}
\format{
A data frame with 104 rows and 8 columns.
}
\source{
Randomised data
}
\description{
A dataset containing raw MFI values and metadata from a sample plate run (Bioplex).
}
\details{
This file is stored in inst/extdata
}
