% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{example_MAGPIX_pk_10std_plate1_csv}
\alias{example_MAGPIX_pk_10std_plate1_csv}
\title{Example Serological Dataset: MAGPIX Plate 1 Pk Analysis 10-Point Standard Curve
A dataset containing raw MFI values and metadata from a sample plate run (MAGPIX).}
\format{
A data frame with 614 rows and 17 columns.
}
\source{
Randomised data
}
\description{
This file is stored in inst/extdata
}
