% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{example_MAGPIX_plate2_csv}
\alias{example_MAGPIX_plate2_csv}
\title{Example Serological Dataset: MAGPIX Plate 2}
\format{
A data frame with 614 rows and 17 columns.
}
\source{
Randomised data
}
\description{
A dataset containing raw MFI values and metadata from a sample plate run (MAGPIX).
}
\details{
This file is stored in inst/extdata
}
