% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{example_platelayout_1}
\alias{example_platelayout_1}
\title{96 Well Plate Example Layout}
\format{
A data frame with 9 rows and 13 variables.
\describe{
  \item{Plate}{Contains rows labelled "A" to "H"}
  \item{1-12}{Contains columns labelled "1" to "12"}
}
}
\description{
96 well plate map in a wide format used in the lab. Contains information of actual Sample ID names in each well.
}
\details{
This file is stored in inst/extdata
}
