% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{excluding_LF016_threshold_values_csv}
\alias{excluding_LF016_threshold_values_csv}
\title{Random Forest Threshold Values for PvSeroTaT Model w/o LF016}
\format{
A data frame with 7 rows and 2 columns.
}
\source{
Mueller Lab, WEHI
}
\description{
Random forest classification model threshold values for maximised sensitivity and specificity,
85, 90, 95% specificity and 85, 90, 95% sensitivity for the PvSeroTaT model with seven antigens.
}
\details{
This file is stored in inst/extdata
}
