% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_validate.R
\name{generate_obj_id}
\alias{generate_obj_id}
\title{Generate object ID}
\usage{
generate_obj_id(digits)
}
\arguments{
\item{digits}{Number of digits of the generated random ID.}
}
\value{
A random number
}
\description{
Generate object ID
}
\examples{
set.seed(4567)
generate_obj_id(5)
}
