% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGithubRelease.R
\name{getGithubRelease}
\alias{getGithubRelease}
\title{Get GitHub Version}
\usage{
getGithubRelease(repo_owner, repo_name)
}
\arguments{
\item{repo_owner}{GitHub Username}

\item{repo_name}{GitHub Repository Name}
}
\value{
Version tag string
}
\description{
A short function to obtain the github version from the repository. This is
a generalisable function that can be used for any version tags on a repo.
}
\author{
Dionne Argyropoulos
}
