% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{longitudinal_MFI_csv}
\alias{longitudinal_MFI_csv}
\title{Training Data Median Fluorescent Intensity (MFI)}
\format{
A data frame with 21,080 rows and 3 columns.
\describe{
  \item{Sample}{Unique identifier for the sample}
  \item{MFI}{Median Fluorescent Intensity}
  \item{Antigen}{Antigen name}
}
}
\source{
Mueller Lab, WEHI
}
\description{
Metadata for the MFI of the training dataset across all eight antigens used
in the PvSeroTaT method.
}
\details{
This file is stored in inst/extdata
}
