% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{longitudinal_RAU_csv}
\alias{longitudinal_RAU_csv}
\title{Training Data Relative Antibody Units (RAU)}
\format{
A data frame with 21,080 rows and 3 columns.
\describe{
  \item{Sample}{Unique identifier for the sample}
  \item{RAU}{Relative Antibody Units}
  \item{Antigen}{Antigen name}
}
}
\source{
Mueller Lab, WEHI
}
\description{
Metadata for the RAU of the training dataset across all eight antigens used
in the PvSeroTaT method.
}
\details{
This file is stored in inst/extdata
}
