% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_cols.R
\name{map_cols}
\alias{map_cols}
\title{Map columns of data}
\usage{
map_cols(
  data,
  dataset_id = NULL,
  id = NULL,
  age_group,
  sex,
  country,
  state = NULL,
  county = NULL,
  city = NULL,
  collection_start_date,
  collection_end_date,
  test_id = NULL,
  result,
  result_cat,
  include_others = FALSE
)
}
\arguments{
\item{data}{a data.frame}

\item{dataset_id}{column, data collection event}

\item{id}{column for IDs}

\item{age_group}{column containing age min and age max for each age group}

\item{sex}{column for sex}

\item{country}{a string, name of the country}

\item{state}{column for state}

\item{county}{column for county}

\item{city}{column for city}

\item{collection_start_date}{column for sampling start date}

\item{collection_end_date}{column for sampling end date}

\item{test_id}{Will add the list of test later}

\item{result}{column with test results}

\item{result_cat}{column with values of "positive", "negative", or "borderline"}

\item{include_others}{include additional columns or not}
}
\value{
A cleaned data.frame
}
\description{
Maps the required columns in your data and validates them.
}
\examples{
library(dplyr)
library(stringr)
df_orig <- data.frame(dataset_id = 1,
                      id = 1:3,
                      age = c("0 - 10", "20-30", "30,40"),
                      sex = c("f", "Male", "binary"),
                      collection_start_date = "2020-03-01",
                      collection_end_date = "2020-04-01",
                      result = c("'2.4'", "3.5", "9"),
                      result_cat = rep("negative", times = 3))

map_cols(df_orig,
         dataset_id = dataset_id,
         id = id,
         age_group = age,
         sex = sex,
         country = "Earth!",
         collection_start_date = collection_start_date,
         collection_end_date = collection_end_date,
         result = result,
         result_cat = result_cat,
         include_others = FALSE)

}
