% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFItoRAU_Adj.R
\name{MFItoRAU_Adj}
\alias{MFItoRAU_Adj}
\title{Median Fluorescent Intensity (MFI) to Relative Antibody Units (RAU)
conversion based on other standard}
\usage{
MFItoRAU_Adj(sero_data, plate_list, qc_results)
}
\arguments{
\item{sero_data}{Output from `readSeroData()`.}

\item{plate_list}{Output from `readPlateLayout()`.}

\item{qc_results}{Output from `runQC()`.}

\item{location}{"PNG" or "ETH" to filter WEHI standard curve data.}
}
\value{
A list of three data frames:
1. Data frame with  MFI data, converted RAU data and matched SampleID's.
2. Plot information for `plotModel` function.
3. Data frame of RAU data for random forest classification use.
}
\description{
This function fits a 5-parameter logistic standard curve to the dilutions
of the positive controls for each protein and converts the MFI values
into relative antibody units (RAU) written by Eamon Conway.
}
\examples{
\donttest{

# Step 0: Load example raw data
your_raw_data <- c(
  system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate2.csv", package = "SeroTrackR")
)
your_plate_layout <- system.file(
  "extdata",
  "example_platelayout_1.xlsx",
  package = "SeroTrackR"
)

# Step 1: Read serology data and plate layout
sero_data  <- readSeroData(your_raw_data,"magpix")
plate_list <- readPlateLayout(your_plate_layout, sero_data)

# Step 2: Process counts and perform quality control
qc_results  <- runQC(sero_data, plate_list)

# Step 3: Convert MFI to RAU using ETH beads
mfi_to_rau <- MFItoRAU_Adj(
  sero_data    = sero_data,
  plate_list   = plate_list,
  qc_results   = qc_results
)

}
}
\author{
Eamon Conway, Dionne Argyropoulos
}
