% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFItoRAU_ETH.R
\name{MFItoRAU_ETH}
\alias{MFItoRAU_ETH}
\title{Median Fluorescent Intensity (MFI) to Relative Antibody Units (RAU)
conversion based on ETH standard}
\usage{
MFItoRAU_ETH(serodata_output, plate_list, counts_QC_output)
}
\arguments{
\item{serodata_output}{Output from `readSeroData()` (reactive).}

\item{plate_list}{Output from `readPlateLayout()` (reactive).}

\item{counts_QC_output}{Output from `getCountsQC()` (reactive).}
}
\value{
A list of three data frames:
1. Data frame with  MFI data, converted RAU data and matched SampleID's.
2. Plot information for `plotModel` function.
3. Data frame of RAU data for random forest classification use.
}
\description{
This function fits a 5-parameter logistic standard curve to the dilutions
of the positive controls for each protein and converts the MFI values
into relative antibody units (RAU) written by Eamon Conway.
}
\author{
Eamon Conway, Dionne Argyropoulos
}
