% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFItoRAU_LDH.R
\name{MFItoRAU_LDH}
\alias{MFItoRAU_LDH}
\title{Median Fluorescent Intensity (MFI) to Relative Antibody Units (RAU)
conversion for LDH}
\usage{
MFItoRAU_LDH(
  sero_data,
  plate_list,
  dilution = c(1e+06, 333333.33, 111111.11, 37037.04, 12345.68, 4115.23, 1371.74, 457.25,
    152.42, 50.81),
  file_path = NULL
)
}
\arguments{
\item{sero_data}{Output from `readSeroData()` or `readSeroData()` (reactive).}

\item{plate_list}{Output from `readPlateLayout()` (reactive).}

\item{dilution}{A list of numbers ranging from S1 to S10. Default: 1000000, 333333.33, 111111.11, 37037.04, 12345.68, 4115.23, 1371.74, 457.25, 152.42, 50.81.}

\item{file_path}{A file path to write the .csv final file. Default: Current working directory.}
}
\value{
A data frame containing the MFI and RAU Dilution values for each sample
}
\description{
This function fits a 5-parameter logistic standard curve to the dilutions
of the positive controls for each protein and converts the MFI values
into relative antibody units (RAU).
}
\examples{
\donttest{
# Example demonstrating multi-plate processing workflow.
# These files are included in the SeroTrackR package under inst/extdata.

your_raw_data <- system.file("extdata", "example_BioPlex_PvLDH_plate1.xlsx", package = "SeroTrackR")

your_plate_layout <- system.file(
  "extdata", "example_platelayout_1.xlsx",
  package = "SeroTrackR"
)

# Read in raw BioPlex data
sero_data <- readSeroData(
  raw_data = your_raw_data,
  platform = "bioplex"
)

# Read matching plate layout
plate_list <- readPlateLayout(
  plate_layout = your_plate_layout,
  sero_data = sero_data
)

# Run MFI to RAU conversion
mfi_outputs <- MFItoRAU_LDH(
  sero_data = sero_data,
  plate_list = plate_list
)

# View All Outputs
mfi_outputs
}
}
\author{
Connie Li Wai Suen, Caitlin Bourke, Dionne Argyropoulos
}
