% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_validate.R
\name{msg_colNotFound}
\alias{msg_colNotFound}
\title{Message when a column was not found}
\usage{
msg_colNotFound(arg, data_name, col_name, ...)
}
\arguments{
\item{arg}{Preset column name or argument for which the validation is run.}

\item{data_name}{Name of the raw data object supplied by user. Can be
supplied using \code{deparse(substitute(data))}}

\item{col_name}{Column name supplied by user for this \code{arg}. Can be supplied
using \code{deparse(substitute(argument))}.}

\item{...}{Additional argument passed to \code{msg_result()}.}
}
\value{
Prints a message to the console.
}
\description{
This is a wrapper around \code{msg_result()}. Only to be used inside
another function.
}
