% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_validate.R
\name{msg_progress}
\alias{msg_progress}
\title{Print message as progress step}
\usage{
msg_progress(msg)
}
\arguments{
\item{msg}{A string scalar.}
}
\value{
A message in the console.
}
\description{
Print message as progress step
}
\examples{
msg_progress("age_group")
}
