% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_validate.R
\name{msg_result}
\alias{msg_result}
\title{Style and print a message}
\usage{
msg_result(arg, msg, error = TRUE, start_time = NULL)
}
\arguments{
\item{arg}{Column name or argument to which the message belongs. This section
will be highlighted in green or red, based on the \code{error} argument.}

\item{msg}{A character vector for the body of the massage(s) that is/are
printed after \code{arg}. This section will be dimmed if \code{error} is set as FALSE.}

\item{error}{If TRUE (default), message starts with a red cross symbol and
\code{arg} will be highlighted in red. If FALSE, message starts with a green
tick symbol and \code{arg} will be highlighted in green.}

\item{start_time}{A POSIXct object used to add code execution time to the
end of the printed message.}
}
\value{
A styled message. If the length of \code{msg} is more than 1, the message
will be broken down in bullet points.
}
\description{
Style and print a message
}
\examples{
timestamp <- Sys.time()
msg_result("age_group", "is a valid column.", error = FALSE,
           start_time = timestamp)
}
