% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_validate.R
\name{msg_wrongType}
\alias{msg_wrongType}
\title{Message when supplied type is wrong}
\usage{
msg_wrongType(
  arg,
  req_type = c("char", "num", "date_char"),
  mode = c("col", "val", "no_typeof"),
  name = NULL,
  data = NULL,
  ...
)
}
\arguments{
\item{arg}{Preset column name or argument for which the validation is run.}

\item{req_type}{Type that was expected for this \code{arg}.}

\item{mode}{Defines if validation is done for a column (col) or for a single
value scalar (val). This will determine how the current wrong type of the
supplied column or vector should be evaluated. The \code{no_typeof} option is for
when we do not need the message to print the type of the supplied input.}

\item{name}{Only for modes \code{col} and \code{val}. The column name, supplied as
\code{deparse(substitute(argument))} or input value as \code{substitute(argument)}.}

\item{data}{A dataframe. Only for mode \code{col}.}

\item{...}{Additional argument passed to \code{msg_result()}.}
}
\value{
Prints a message to the console.
}
\description{
Message when supplied type is wrong
}
\examples{
msg_wrongType("age", req_type = "num", mode = "no_typeof")
}
