% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{PkPfPv_Panel_1}
\alias{PkPfPv_Panel_1}
\title{Pk/Pf/Pv Panel Antigens and Species List}
\format{
A data frame with 17 rows and 2 columns.
}
\source{
Longley Lab, WEHI
}
\description{
A list containing the Pk/Pf/Pv antigens names and associated species. In the future there will be multiple
options, with this list here serving as the first list.
}
\details{
This file is stored in inst/extdata
}
