% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{platemap_csv}
\alias{platemap_csv}
\title{96 Well Plate Map}
\format{
A data frame with 96 rows and 2 columns.
\describe{
  \item{Location}{Location on plate}
  \item{Well}{Well on plate}
}
}
\description{
96 well plate map in a long format used for data wrangling.
}
\details{
This file is stored in inst/extdata
}
