% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeadCounts.R
\name{plotBeadCounts}
\alias{plotBeadCounts}
\title{Plot Bead Counts per Plate per Antigen}
\usage{
plotBeadCounts(antigen_counts_output)
}
\arguments{
\item{antigen_counts_output}{Output from `getAntigenCounts()` (reactive).}
}
\value{
Dot plot with values > 15 threshold coloured in blue (sufficient
beads) and less than or equal to 15 beads coloured in red (repeat)
faceted by each antigen
(ggplot).
}
\description{
Enhances the `plotCounts()` output by providing greater resolution,
displaying antigens per plate, and enabling SampleID name visibility via
hover (transformed to Plotly in server.R)
}
\examples{
\donttest{

# Step 0: Load example raw data
your_raw_data <- c(
  system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate2.csv", package = "SeroTrackR")
)
your_plate_layout <- system.file(
  "extdata",
  "example_platelayout_1.xlsx",
  package = "SeroTrackR"
)

# Step 1: Read serology data and plate layout
sero_data  <- readSeroData(your_raw_data,"magpix")
plate_list <- readPlateLayout(your_plate_layout, sero_data)

# Step 2: Process counts and perform quality control
counts      <- processCounts(sero_data)
counts_raw  <- getCounts(counts)
sample_ids  <- getSampleID(counts, plate_list)
antigen_cts <- getAntigenCounts(counts, plate_list)

# Step 3: Plot Bead Counts
plotBeadCounts(antigen_cts)
}
}
\author{
Dionne Argyropoulos
}
