% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxPlotClassification.R
\name{plotBoxPlotClassification}
\alias{plotBoxPlotClassification}
\title{Plot Classification}
\usage{
plotBoxPlotClassification(all_classifications, selected_threshold)
}
\arguments{
\item{all_classifications}{Data frame of `classifyResults()` for all
sens_spec thresholds.}

\item{selected_threshold}{String with the threshold (reactive).}
}
\value{
Box plots with RAU values for each protein stratified by
classification (ggplot).
}
\description{
One example of data visualisation to detect the median and interquartile
range of the RAU values per antigen for seropositive and seronegative
individuals. Please note that the `classifyResults()` function must
be run first.
}
\author{
Dionne Argyropoulos
}
