% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotManyStdCurves.R
\name{plotManyStdCurves}
\alias{plotManyStdCurves}
\title{Standard Curve for multiple plates}
\usage{
plotManyStdCurves(runPlasmoSeropoint_output)
}
\arguments{
\item{runPlasmoSeropoint_output}{Output from `runPlasmoSero5point()` or `runPlasmoSero10point()`s}
}
\value{
ggplot of standard curves for each protein and species
}
\description{
This function is primarily for the species Pk/Pf/Pv analysis to compare and visualise the MFI of standards.
}
\author{
Dionne Argyropoulos
}
