% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOneStdCurve.R
\name{plotOneStdCurve}
\alias{plotOneStdCurve}
\title{Standard Curve for one plate}
\usage{
plotOneStdCurve(runPlasmoSeropoint_output, plate_name)
}
\arguments{
\item{runPlasmoSeropoint_output}{Output from `runPlasmoSero5point()` or `runPlasmoSero10point()`}

\item{plate_name}{The plate you would like to visualise}
}
\value{
ggplot of standard curves for each protein and species
}
\description{
This function is primarily for the species Pk/Pf/Pv analysis to visualise the MFI of standards.
}
\author{
Dionne Argyropoulos
}
