% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStds.R
\name{plotStds}
\alias{plotStds}
\title{Plot Raw Median Fluorescent Intensity of Standard Curve Data}
\usage{
plotStds(sero_data, location, experiment_name)
}
\arguments{
\item{sero_data}{Output from `readSeroData()` (reactive).}

\item{location}{"PNG" or "ETH" to filter WEHI standard curve data (reactive).}

\item{experiment_name}{User-input experiment name (reactive).}
}
\value{
- Dot and line plot of standard curves (S1-S10) with PNG or Ethiopia stds
underneath (ggplot).
- WEHI-acceptable standard curve data on background of plot with user data.
}
\description{
This function gets the standards data and plots the standard curves.
}
\examples{
\donttest{
# Example demonstrating how to process bead count data.
# These files are included in the SeroTrackR package under inst/extdata.

your_raw_data <- c(
  system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate2.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate3.csv", package = "SeroTrackR")
)

# Read in raw MAGPIX data
sero_data <- readSeroData(
  raw_data = your_raw_data,
  platform = "magpix"
)

# Plot Standards
plotStds(
  sero_data = sero_data,
  location = "ETH",
  experiment_name = "experiment1"
)

}

}
\author{
Shazia Ruybal-Pesantez, Dionne Argyropoulos
}
