% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStds_all.R
\name{plotStds_all}
\alias{plotStds_all}
\title{Plot Raw Median Fluorescent Intensity of Standard Curve Data}
\usage{
plotStds_all(serodata_output, experiment_name)
}
\arguments{
\item{serodata_output}{Output from `readSeroData()` (reactive).}

\item{experiment_name}{User-input experiment name (reactive).}
}
\value{
- Dot and line plot of standard curves (S1-S10)
- WEHI-acceptable standard curve data on background of plot with user data.
}
\description{
This function gets the standards data and plots the standard curves for any antigens (i.e., non-PvSeroTaT specific).
}
\author{
Shazia Ruybal-Pesantez, Dionne Argyropoulos
}
