% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{png_eth_stds_csv}
\alias{png_eth_stds_csv}
\title{Plate Metadata}
\format{
A data frame with 710 rows and 6 columns.
\describe{
  \item{std_plate}{Unique plate with S1-10}
  \item{Sample}{Unique identifier for the sample}
  \item{antigen}{Antigen name}
  \item{eth_mfi}{Median Fluorescent Intensity for Ethiopian IgG coupled beads}
  \item{dilution}{Standard serial dilution factor}
  \item{png_mfi}{Median Fluorescent Intensity for PNG IgG coupled beads}
}
}
\source{
Mueller Lab, WEHI
}
\description{
MFI values from QA/QC plate with Papua New Guinean and Ethiopian IgG coupled beads
for a range of the eight antigens in the PvSeroTaT method.
}
\details{
This file is stored in inst/extdata
}
