% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPkPfPv.R
\name{processPkPfPv}
\alias{processPkPfPv}
\title{Processing Serological Data for Pk/Pf/Pv MFI to RAU conversion}
\usage{
processPkPfPv(serodata_output, plate_list, panel = "panel1")
}
\arguments{
\item{serodata_output}{Output of `readSeroData()`}

\item{plate_list}{Output of `readPlateLayout()`}

\item{panel}{Panel of Pk/Pf/Pv antigens. Default = "panel1".}
}
\value{
A list of two data frames:
1. Data frame with Pk antigens
2. Data frame with Pf/Pv antigens
}
\description{
This is a pre-requisite function before running the `MFItoRAU_Plasmo()` so that the
appropriate MFI to RAU conversions can be run for the respective antigens.
}
\author{
Dionne Argyropoulos
}
