% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvsero-package.R
\docType{package}
\name{pvsero-package}
\alias{pvsero}
\alias{pvsero-package}
\title{pvsero: Functions for the Sero Analysis}
\description{
Data wrangling and cleaning, quality control checks and implementation of machine learning classification algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dionnecargy/pvsero}
  \item Report bugs at \url{https://github.com/dionnecargy/pvsero/issues}
}

}
\author{
\strong{Maintainer}: Dionne Argyropoulos \email{argyropoulos.d@wehi.edu.au}

}
\keyword{internal}
