% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{PvSeroTaTmodel_rds}
\alias{PvSeroTaTmodel_rds}
\title{Eight-Antigen Random Forest Classification Model}
\format{
An Rdata file containing multiple R objects.
}
\source{
Mueller Lab, WEHI
}
\description{
RData containing the PvSeroTaT eight antigen random forest classification model with
10-fold cross-validation and 5 repeats, 10,000 trees and trained on the dataset described in
Smith et al., in prep.
}
\details{
This file is stored in inst/extdata
}
