% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{random_forest_excludingLF016_rds}
\alias{random_forest_excludingLF016_rds}
\title{Seven-Antigen Random Forest Classification Model w/o LF016}
\format{
An Rdata file containing multiple R objects.
}
\source{
Mueller Lab, WEHI
}
\description{
RData containing the PvSeroTaT seven antigen random forest classification model (without LF016) with
10-fold cross-validation and 5 repeats, 10,000 trees and trained on the dataset described in
Smith et al., in prep.
}
\details{
This file is stored in inst/extdata
}
