% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{regions}
\alias{regions}
\title{Regions list}
\format{
\code{regions}
A list. Each element of the list returns a unique ID.
}
\source{
Based on the CGAZ dataset from \href{https://github.com/wmgeolab/geoBoundaries/tree/main/releaseData/CGAZ}{geoBoundaries}.
}
\usage{
regions
}
\description{
A nested named list of regions at three administrative (ADM) levels. This
list can be used in the \code{st_validate()} or \code{st_locate()} functions to
rename geographic areas based on SeroTracker's predefined list of region
names. This will help automate the generation of visualizations and
analyses.
}
\details{
adm0: Use this level to select your study's country.

adm1: Use this level to select your study's state/province.

adm2: Use this level to select your study's district/municipality or equivalent division.
}
\examples{
regions$adm2$Canada$Alberta$Calgary
}
\keyword{datasets}
