% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeroData.R
\name{relabel_columns}
\alias{relabel_columns}
\title{Relabel column names to Standardised Naming Convention}
\usage{
relabel_columns(df)
}
\arguments{
\item{df}{Data frame from `readSeroData()`.}
}
\value{
A data fame with columns renamed
}
\description{
This is a helper function to be used inside `readSeroData()` to relabel
columns for each plate.
}
\author{
Dionne Argyropoulos
}
