% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLDHPipeline.R
\name{runLDHPipeline}
\alias{runLDHPipeline}
\title{Run LDH Pipeline from Start to End}
\usage{
runLDHPipeline(
  raw_data,
  plate_layout,
  platform = "bioplex",
  location = "ETH",
  dilution = c(1e+06, 333333.33, 111111.11, 37037.04, 12345.68, 4115.23, 1371.74, 457.25,
    152.42, 50.81),
  experiment_name = "experiment1",
  file_path = NULL
)
}
\arguments{
\item{raw_data}{String with the raw data path.}

\item{plate_layout}{An ".xlsx" file with sheets labelled plate1, plate2... etc.}

\item{platform}{"magpix" or "bioplex". Default: "Bioplex"}

\item{location}{"PNG" or "ETH" to filter WEHI standard curve data. Default: "ETH"/}

\item{dilution}{A list of numbers ranging from S1 to S10. Default: 1000000, 333333.33, 111111.11, 37037.04, 12345.68, 4115.23, 1371.74, 457.25, 152.42, 50.81.}

\item{experiment_name}{User-input experiment name. Default: "experiment1".}

\item{file_path}{A file path to write the .csv final file. Default: Current working directory.}
}
\value{
A data frame containing the MFI and RAU Dilution values for each sample, QC plots for standard curve, bead counts and blanks.
}
\description{
A master function combining the entire LDH pipeline into one command to run in R.
}
\author{
Dionne Argyropoulos
}
