% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPlasmoSero5point.R
\name{runPlasmoSero5point}
\alias{runPlasmoSero5point}
\title{Read Serological Data and 5-point Standard Curve}
\usage{
runPlasmoSero5point(
  raw_data,
  platform,
  plate_layout,
  date = format(Sys.Date(), "\%Y\%m\%d")
)
}
\arguments{
\item{raw_data}{String with the raw data path.}

\item{platform}{"magpix" or "bioplex".}

\item{plate_layout}{File is labelled with S1-5 with ETH as a suffix, S1 to S5 with PK suffix, Blanks and NC.}

\item{date}{String with today's date}
}
\value{
A data frame containing all samples and their RAU values.
}
\description{
A function for generating RAU values with the two curves currently being used in the Pk analysis assay.
This is reliant on a 5-fold 5-point serial dilution covering S1 to S5 for poth Pk pool and ETH pool.
}
\author{
Dionne Argyropoulos, Caitlin Bourke
}
