% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQC.R
\name{runQC}
\alias{runQC}
\title{Run Quality Control Pipeline}
\usage{
runQC(sero_data, plate_list)
}
\arguments{
\item{sero_data}{Output from `readSeroData()`.}

\item{plate_list}{Output from `readPlateLayout()`.}
}
\value{
processCounts_output, getCounts_output, sampleid_output, getAntigenCounts_output, getCountsQC_output
}
\description{
A master function containing each quality control processing step.
}
\examples{

\donttest{
# Example data supplied with the package
your_raw_data <- c(
  system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate2.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate3.csv", package = "SeroTrackR")
)

plate_layout <- system.file(
  "extdata", "example_platelayout_1.xlsx", package = "SeroTrackR"
)

# Read serology data and plate layout
sero_data  <- readSeroData(your_raw_data,"magpix")
plate_list <- readPlateLayout(your_plate_layout, sero_data)

# Run full pipeline including classification
runQC(
  sero_data = sero_data,
  plate_list = plate_list
)
}
}
\author{
Dionne Argyropoulos
}
