% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSeroPipeline.R
\name{runSeroPipeline}
\alias{runSeroPipeline}
\title{Run Serology Analysis Pipeline from Start to End}
\usage{
runSeroPipeline(
  raw_data,
  raw_data_filenames,
  platform,
  location,
  experiment_name
)
}
\arguments{
\item{raw_data}{String with the raw data path (reactive).}

\item{raw_data_filenames}{String with the raw data filenames (reactive).}

\item{platform}{"magpix" or "bioplex" (reactive).}

\item{location}{"PNG" or "ETH" to filter WEHI standard curve data (reactive).}

\item{experiment_name}{User-input experiment name (reactive).}
}
\value{
mfi_to_rau_output, stdcurve_plot, plateqc_plot, check_repeats_output, blanks_plot, model_plot
}
\description{
A master function combining the entire Serology Analysis pipeline into one command to run in R.
This is a similar function as `runPvSeroPipeline()` except can be used for non-Pv samples as there is no classification step.
}
\author{
Dionne Argyropoulos
}
