% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_raw_data.R
\docType{data}
\name{sample_raw_data}
\alias{sample_raw_data}
\title{Sample raw data}
\format{
\code{sample_raw_data}
A data frame with 3 rows and 10 columns:
\describe{
\item{dataset_id}{ID of each collection period}
\item{id}{Anonimized ID of each participant or sample}
\item{age}{Age}
\item{sex}{Sex}
\item{state}{Administative level 1 region names}
\item{city}{Administative level 2 region names}
\item{collection_start_date}{Collection start date}
\item{collection_end_date}{Collection end date}
\item{result}{Test results}
\item{result_cat}{Interpretation of test results; e.g. positive, negative, or borderline}
}
}
\source{
Fictitious data
}
\usage{
sample_raw_data
}
\description{
Example dataset to showcase the package's expectations from a raw dataset and the package's workflow to make the raw data ready for submission to SeroTracker 2.0.
}
\keyword{datasets}
