% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_xlsx.R
\name{save_xlsx}
\alias{save_xlsx}
\title{Save as xlsx}
\usage{
save_xlsx(data, path)
}
\arguments{
\item{data}{a validated data.frame, output of st_validate()}

\item{path}{where to save the xlsx file}
}
\value{
Saves data.fram to disk
}
\description{
Save as xlsx
}
\examples{
validated_df <- st_validate(
  sample_raw_data,
  dataset_id = dataset_id,
  id = id,
  age_group = "12-17",
  sex = "m",
  adm0 = regions$adm0$Canada,
  adm1 = regions$adm1$Canada$Alberta,
  adm2 = regions$adm2$Canada$Alberta$Calgary,
  collection_start_date = "2023-01-01",
  collection_end_date = "2023-02-01",
  test_id = assays$`SARS-CoV-2`$`AAZ LMB - IgG, IgM - COVID-PRESTO®`,
  result = result,
  result_cat = "negative",
  include_others = TRUE
)

save_xlsx(validated_df, path = tempdir())

}
