% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_locate.R
\name{st_locate}
\alias{st_locate}
\title{Convert region names to codes}
\usage{
st_locate(
  data,
  adm0,
  adm1,
  adm2 = NULL,
  ...,
  into = c("adm1", "adm2"),
  n_unmatched_printed = 20
)
}
\arguments{
\item{data}{A dataframe.}

\item{adm0}{A string representing one country (adm0) code. Use
\code{serotrackr::regions$adm0$YourCountry} to select it.}

\item{adm1, adm2}{a string or an unquoted name of a character column that
contains the state/province (adm1), or district/municipality
(adm2) names or codes. If your study is conducted in only one adm1 or adm2
region, use \code{serotrackr::regions} to select them.}

\item{...}{A sequence of two-sided assignments as in
\code{"region_name" = region_code}. \code{region_name} is your unmatched region name.
\code{region_code} must be in the form of \code{regions$adm1$YourCountry$YourState}
for unmatched adm1 regions and
\code{regions$adm2$YourCountry$YourState$YourDistrict} for unmatched adm2
regions. \code{st_locate()} uses \code{dplyr::case_when()} internally for this part.}

\item{into}{A character vector of length 1 or 2, specifying the name(s) of
the new columns that are added to your data for adm1 (and adm2) region
codes. If length is one, name of the adm1 column will be assigned. If
length is two, names of adm1 and adm2 columns will be assigned,
respectively.}

\item{n_unmatched_printed}{A single number, indicating the quantity of
unmatched unique region names that are printed to the console. This argument
is implemented to prevent potentially flooding your console with unmatched
region names. The default value is 20. If there are more than 20 unmatched
region names in your data, increase this number to see the rest.}
}
\value{
A data.frame that consists of the input data.frame plus one (or two)
new columns containing adm1 (and adm2) region codes. You can then use these
two new columns for adm1 and/or adm2 arguments of \code{st_validate()}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function helps you convert your region names to region codes acceptable
by \code{serotrackr}. It is usually used when you get an error for
either \code{adm1} or \code{adm2} arguments in \code{st_validate()}. In that case, you
can use this function to convert your region names to codes and make sure
they pass the validation by \code{st_validate()}.

\code{st_locate()} uses its first three (or four) arguments to automatically
match your region names with \code{serotrackr}'s predefined region codes, which
are stored in \code{serotrackr::regions_df}. It performs exact case insensitive
matching and also ignores accented letters. It then produces a report of
your region names that could not be matched. You can then use
\code{st_locate()}'s ... (ellipsis) argument to define those unmatched region
names.
}
\examples{
st_locate(
  data = sample_raw_data,
  adm0 = regions$adm0$Canada,
  adm1 = state,
  adm2 = city,
  "Toronoto" = regions$adm2$Canada$Ontario$Toronto,
  "Calagry"  = regions$adm2$Canada$Alberta$Calgary
)

}
