% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_save.R
\name{st_save}
\alias{st_save}
\title{Export data to Excel}
\usage{
st_save(data, estimates, path)
}
\arguments{
\item{data}{A validated data.frame, output of \code{st_validate()}}

\item{estimates}{a data.frame, output of \code{st_aggregate()}}

\item{path}{where to save the xlsx file}
}
\value{
An xlsx file saved to disc
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Organizes individual level data and aggregate estimates into a single xlsx
file to be uploaded on the \href{https://serotracker.com}{SeroTracker}
website.
}
\examples{
mydata <- dplyr::mutate(
  sample_raw_data,
  age = ifelse(age \%in\% c(-999, 999), NA, age)
)

validated_df <- st_validate(
  mydata,
  dataset_id = dataset_id,
  id = id,
  age_group = age_group,
  age = age,
  sex = sex,
  adm0 = regions$adm0$Canada,
  adm1 = regions$adm1$Canada$Alberta,
  adm2 = regions$adm2$Canada$Alberta$Calgary,
  collection_start_date = "2020-Mar-01",
  collection_end_date = "15/8/2023",
  test_id = assays$`SARS-CoV-2`$`ID.Vet - IgG - ID Screen`,
  result = result,
  result_cat = result_cat,
  include_others = TRUE,
  rmd_safe = TRUE
)

estimates <- st_aggregate(validated_df)

st_save(validated_df, estimates, path = tempdir())

}
