% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_validate.R
\name{st_validate}
\alias{st_validate}
\title{Validate individual level data}
\usage{
st_validate(
  data,
  dataset_id,
  id,
  adm0,
  adm1 = NULL,
  adm2 = NULL,
  collection_start_date,
  collection_end_date,
  test_id,
  result,
  result_cat = NULL,
  age_group = NULL,
  age = NULL,
  sex = NULL,
  include_others = TRUE
)
}
\arguments{
\item{data}{a data.frame.}

\item{dataset_id}{An unquoted column name or a length-one vector that
differentiates the data collection event(s).}

\item{id}{column for anonymized individual level IDs. This column will be
used to generate aggregate estimates.}

\item{adm0, adm1, adm2}{a string or an unquoted name of a character column that
contains the country (adm0), state/province (adm1), or district/municipality
(adm2) codes. Use serotrackr::regions to select these. Only one adm0 is
acceptable. adm1 and adm2 can be more.}

\item{collection_start_date, collection_end_date}{Unquoted name of a date or
character column or a date or string scalar (vector of length one) for
sampling start and end dates. \code{lubridate::parse_date_time2()} is used to
parse dates. Only \code{yyyy-mm-dd} or \code{dd-mm-yyyy} structures are acceptable.
It recognize arbitrary non-digit separators as well as no separator. Month
Can be entered as a digit or a full or abbreviated name.}

\item{test_id}{a string or an unquoted name of a character column that
contains the test IDs. Use serotrackr::assays to select these.}

\item{result}{Unquoted name of a numeric column containing test results.}

\item{result_cat}{Unquoted name of a character column with values of
\code{positive}, \code{borderline}, or \code{negative}, ignoring case. A single string
is also acceptable.}

\item{age_group}{Unquoted name of a character column or a string containing
age group(s). The only structures acceptable are \code{number-number} or
\verb{number+}. E.g. 18-64, and 65+.}

\item{age}{Unquoted name of a numeric column or a single number. Acceptable
values are between 0 and 120 inclusive.}

\item{sex}{Unquoted name of a character column or a string. Acceptable
values are: \code{f}, \code{m}, \code{o}, \code{female}, \code{male}, or \code{other} ignoring case.}

\item{include_others}{include additional columns or not}
}
\value{
A validated data.frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Locates the required and optional columns in your data and validates them.
}
\examples{
st_validate(sample_raw_data,
            dataset_id = dataset_id,
            id = id,
            age_group = "12-17",
            sex = "m",
            adm0 = regions$adm0$Canada,
            adm1 = regions$adm1$Canada$Alberta,
            adm2 = regions$adm2$Canada$Alberta$Calgary,
            collection_start_date = "2023-01-01",
            collection_end_date = "2023-02-01",
            test_id = assays$`SARS-CoV-2`$`AAZ LMB - IgG, IgM - COVID-PRESTO®`,
            result = result,
            result_cat = "negative",
            include_others = TRUE)
}
