% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_validate.R
\name{validate_data}
\alias{validate_data}
\title{Validate using predefined rules}
\usage{
validate_data(
  data,
  rules,
  mode = c("col", "val"),
  col_to_validate = names(data),
  rules_msgCut = NULL
)
}
\arguments{
\item{data}{A dataframe whose colnames are exactly those specified in rules}

\item{rules}{A validator object from the \code{validate} package}

\item{mode}{Defines if validation is done for a column (col) or for a single
value scalar (val). Th latter has shorter error messages.}

\item{col_to_validate}{Column that is the target of validation. Specify this
when validation of the target column requires other columns from data; e.g
to asses end date is after start date.}

\item{rules_msgCut}{A character vector of rule names for which the last part
of the returned message will be trimmed; e.g. for rule \code{is.infinite()}.}
}
\value{
A character vector of error messages based on rules applied. If no
errors occur, a vector of length zero is returned.
}
\description{
Validate using predefined rules
}
\examples{
err_msgs <- validate_data(sample_raw_data, rules["sex_presetVal"],
                          mode = "col", col_to_validate = "sex")
err_msgs
}
