% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyPv.r
\name{classifyPv}
\alias{classifyPv}
\title{Random Forest Classification for Pv only in Pk/Pf/Pv analysis}
\usage{
classifyPv(mfi_to_rau_output, algorithm_type, sens_spec, qc_results)
}
\arguments{
\item{mfi_to_rau_output}{Output from `MFItoRAU()` or `MFItoRAU_Adj()`.}

\item{algorithm_type}{User-selected algorithm choice:
- "antibody_model" (PvSeroTaT model; default), or
- "antibody_model_excLF016" (PvSeroTaT excluding LF016).}

\item{sens_spec}{User-selected Sensitivity/Specificity threshold:
- "balanced" (default),
- "85\% sensitivity",
- "90\% sensitivity",
- "95\% sensitivity",
- "85\% specificity",
- "90\% specificity".
- "95\% specificity".}

\item{qc_results}{Output from `runQC()`.}
}
\value{
- Data frame with exposure status for every sample.
- Summary table with positive/negative results for each threshold.
}
\description{
This function classifies unknown samples as recently exposed or not
(Note: MFItoRAU() or MFItoRAU_Adj() needs to be run first to convert to
RAU). This is a slightly modified function for ONLY use in Pk/Pf/Pv analysis.
The only difference is that ETHtoPNGloglog_Dilution is used to identify the
correct columns to then classify.
}
\examples{
\donttest{

# Step 0: Load example raw data
raw_files_5std <- c(
  system.file("extdata", "example_MAGPIX_pk_5std_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_pk_5std_plate2.csv", package = "SeroTrackR")
)
plate_layout_5std <- system.file(
  "extdata",
  "example_platelayout_pk_5std.xlsx",
  package = "SeroTrackR"
)

# Step 1: Read in serology data and plate layout
sero_data  <- readSeroData(raw_files_5std, "magpix")
plate_list <- readPlateLayout(plate_layout_5std, sero_data)

# Step 2: Process counts and QC
qc_results  <- runQC(sero_data, plate_list)

# Step 3: Convert MFI to RAU using 5-point standard curve
mfi_results <- MFItoRAU_Plasmo(
  sero_data    = sero_data,
  plate_list   = plate_list,
  panel        = "panel1",
  std_point    = 5,
  qc_results   = qc_results
)

# Step 4: Classify Pv samples
pv_classified <- classifyPv(
  mfi_to_rau_output = mfi_results,
  algorithm_type    = "antibody_model",
  sens_spec         = "balanced",
  qc_results        = qc_results
)
}
}
\author{
Lauren Smith, Dionne Argyropoulos
}
