% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dilution_to_mfi.R
\name{convert_dilution_to_mfi}
\alias{convert_dilution_to_mfi}
\title{Convert known dilution to mfi from fitted standard curve}
\usage{
convert_dilution_to_mfi(dilution, params)
}
\arguments{
\item{dilution}{Known dilution of samples}

\item{params}{Known parameters for five parameter logistic fit.}
}
\value{
Returns the predicted mfi of a sample with known dilution.
}
\description{
Convert dilution to predicted mfi using known standard curve fit.
}
\author{
Eamon Conway
}
