% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeroData.R
\name{.check_platform}
\alias{.check_platform}
\title{Check Platform}
\usage{
.check_platform(raw_data, platform)
}
\arguments{
\item{raw_data}{String with the raw data path.}

\item{platform}{"magpix" or "bioplex".}
}
\value{
TRUE: if platform == file format, ERROR message when platform does
not equal file format.
}
\description{
This function checks the platform the user has input and whether it aligns
with the correct format as expected. Will report error if NOT aligned.
}
\examples{
your_raw_data <- system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR")
.check_platform(raw_data = your_raw_data, platform = "magpix")
}
\author{
Dionne Argyropoulos
}
