% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeroData.R
\name{.clean_bioplex}
\alias{.clean_bioplex}
\title{Helper function to process bioplex data}
\usage{
.clean_bioplex(df)
}
\arguments{
\item{file}{String with the raw data path.}
}
\value{
Cleaned data fame
}
\description{
Helper function to process bioplex data
}
\examples{
your_raw_data <- system.file("extdata", "example_BioPlex_plate1.xlsx", package = "SeroTrackR")
df            <- .read_luminex_file(your_raw_data)
results       <- .clean_bioplex(df)
}
\author{
Dionne Argyropoulos
}
