% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeroData.R
\name{.magpix_version_config}
\alias{.magpix_version_config}
\title{Helper function to identify Magpix version}
\usage{
.magpix_version_config(version)
}
\arguments{
\item{version}{String with the raw data path.}
}
\value{
specific column names for filtering for xPONENT software v4.2 and v4.3
}
\description{
Helper function to identify Magpix version
}
\examples{
version = "4.2"
.magpix_version_config(version)

}
\author{
Dionne Argyropoulos
}
