% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeroData.R
\name{.post_process_bioplex}
\alias{.post_process_bioplex}
\title{Helper function to process bioplex sections}
\usage{
.post_process_bioplex(df)
}
\arguments{
\item{file}{String with the raw data path (reactive).}
}
\value{
List of data_raw, results, counts, blanks, stds, run
}
\description{
Helper function to process bioplex sections
}
\author{
Dionne Argyropoulos

your_raw_data <- system.file("extdata", "example_BioPlex_plate1.xlsx", package = "SeroTrackR")
df            <- .read_luminex_file(your_raw_data)
sections      <- .post_process_bioplex(df)
}
