% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeroData.R
\name{.post_process_luminex}
\alias{.post_process_luminex}
\title{Helper function to process luminex into master_list}
\usage{
.post_process_luminex(sections, file_name, master_list)
}
\arguments{
\item{sections}{Output from `.post_process_bioplex()`.}

\item{file_name}{User input file name.}

\item{master_list}{Intermediary df from `readSeroData()`.}
}
\value{
List of data_raw, results, counts, blanks, stds, run
}
\description{
Helper function to process luminex into master_list
}
\author{
Dionne Argyropoulos
}
