% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_standard_curve.R
\name{fit_standard_curve}
\alias{fit_standard_curve}
\title{Fit a standard curve to known mfi and dilution values.}
\usage{
fit_standard_curve(mfi, dilution, control = NULL)
}
\arguments{
\item{mfi}{Known mfi of samples}

\item{dilution}{Known dilution of samples}

\item{control}{Optional list of control parameters for the underlying call to optim.}
}
\description{
We wish to convert the standard curve samples to a five parameter logistic curve.
This function takes those values and calls optim to determine the fit.
}
\author{
Eamon Conway
}
