% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAntigenCounts.R
\name{getAntigenCounts}
\alias{getAntigenCounts}
\title{Get Count Data for each Antigen from the Raw Median Fluorescent Intensity}
\usage{
getAntigenCounts(processed_counts, plate_list)
}
\arguments{
\item{processed_counts}{Output from `processCounts()` (reactive).}

\item{plate_list}{Output from `readPlateLayout()` (reactive).}
}
\value{
(i) Data frame providing bead counts per antigen per well per plate.
(ii) Designates whether wells should be repeated if there are \eqn{<=}  15 beads
(repeat) or if they are sufficient with > 15 beads (sufficient beads).
}
\description{
This function obtains the count data from the raw Median Fluorescent
Intensity (MFI). This function relies on the `readAntigens` and
`readSeroData` data processing functions.
}
\author{
Dionne Argyropoulos
}
