% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCounts.R
\name{getCounts}
\alias{getCounts}
\title{Get Count Data from Raw Median Fluorescent Intensity}
\usage{
getCounts(processed_counts)
}
\arguments{
\item{processed_counts}{Output from `processCounts()` (reactive).}
}
\value{
(i) Data frame providing bead counts per well per plate. (ii)
Designates whether wells should be repeated if there are ≤ 15 beads (repeat)
or if they are sufficient with > 15 beads (sufficient beads).
}
\description{
This function obtains the count data from the raw Median Fluorescent
Intensity (MFI). This is an interim function used for the plotCounts
function. This function relies on the `readAntigens` and `readSeroData` data
processing functions.
}
\author{
Shazia Ruybal-Pesántez, Dionne Argyropoulos
}
