% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQC.R
\name{getCountsQC}
\alias{getCountsQC}
\title{Get All Counts Data}
\usage{
getCountsQC(antigen_counts_output, counts_output)
}
\arguments{
\item{antigen_counts_output}{Output from `getAntigenCounts`.}

\item{counts_output}{Output from `getCounts`.}
}
\value{
Joined data frame for all count data.
}
\description{
This function obtains the count data from the raw Median Fluorescent
Intensity (MFI). This function relies on the output of the Antigen-specific
counts (`getAntigenCounts`) and the Well or Sample-specific counts
(`getCounts`).
}
\examples{
\donttest{

# Step 0: Load example raw data
your_raw_data <- c(
  system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate2.csv", package = "SeroTrackR")
)
your_plate_layout <- system.file(
  "extdata",
  "example_platelayout_1.xlsx",
  package = "SeroTrackR"
)

# Step 1: Read serology data and plate layout
sero_data  <- readSeroData(your_raw_data,"magpix")
plate_list <- readPlateLayout(your_plate_layout, sero_data)

# Step 2: Process counts and perform quality control
counts      <- processCounts(sero_data)
counts_raw  <- getCounts(counts)
sample_ids  <- getSampleID(counts, plate_list)
antigen_cts <- getAntigenCounts(counts, plate_list)
counts_qc   <- getCountsQC(antigen_cts, counts_raw)

}
}
\author{
Dionne Argyropoulos
}
