% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountsQC.R
\name{getCountsQC}
\alias{getCountsQC}
\title{Get All Counts Data}
\usage{
getCountsQC(antigen_counts_output, counts_output)
}
\arguments{
\item{antigen_counts_output}{Output from `getAntigenCounts` (reactive).}

\item{counts_output}{Output from `getCounts` (reactive).}
}
\value{
Joined data frame for all count data.
}
\description{
This function obtains the count data from the raw Median Fluorescent
Intensity (MFI). This function relies on the output of the Antigen-specific
counts (`getAntigenCounts`) and the Well or Sample-specific counts
(`getCounts`).
}
\author{
author
}
