% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlateLayout.R
\name{getPlateLayout}
\alias{getPlateLayout}
\title{Find and create a master plate layout file}
\usage{
getPlateLayout(folder_path = getwd())
}
\arguments{
\item{folder_path}{A string containing your main folder for your project or the plate layout files. Default = current working directory.}
}
\value{
An .xlsx file saved to your current working directory with multiple tabs, one tab for each plate layout.
}
\description{
Join multiple a plate layout files into one master file with multiple tabs
}
\author{
Dionne Argyropoulos
}
