% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRepeats.R
\name{getRepeats}
\alias{getRepeats}
\title{Check Beads to Repeat}
\usage{
getRepeats(qc_results, plate_list)
}
\arguments{
\item{qc_results}{Output from `runQC()`.}

\item{plate_list}{Output from `readPlateLayout()`.}
}
\value{
A data frame with wells to "fail", OR if no "fail" found will return
text "No repeats necessary".
}
\description{
This function gets the count data and outputs a table of the isolates to
repeat or a statement to confirm that none need to be repeated.
}
\examples{
\donttest{
# Step 0: Load example raw data and plate layout
raw_data <- c(
  system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate2.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate3.csv", package = "SeroTrackR")
)
plate_layout <- system.file("extdata", "example_platelayout_1.xlsx", package = "SeroTrackR")

# Step 1: Read data and plate layout
sero_data   <- readSeroData(raw_data, platform = "magpix")
plate_list  <- readPlateLayout(plate_layout, sero_data)

# Step 2: Process counts
qc_results  <- runQC(sero_data, plate_list)

# Step 3: Identify samples to repeat
repeats_table <- getRepeats(
  qc_results = qc_results,
  plate_list = plate_list
)

# View results
repeats_table
}
}
\author{
Dionne Argyropoulos
}
