% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRepeats.R
\name{getRepeats}
\alias{getRepeats}
\title{Check Beads to Repeat}
\usage{
getRepeats(counts_output, processed_counts, plate_list)
}
\arguments{
\item{counts_output}{Output from `getCounts()` (reactive).}

\item{processed_counts}{Output from `processCounts()`.}

\item{plate_list}{Output from `readPlateLayout()` (reactive).}
}
\value{
A data frame with wells to "fail", OR if no "fail" found will return
text "No repeats necessary".
}
\description{
This function gets the count data and outputs a table of the isolates to
repeat or a statement to confirm that none need to be repeated.
}
\author{
Dionne Argyropoulos
}
