% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleID.R
\name{getSampleID}
\alias{getSampleID}
\title{Get SampleID from Plate Layout}
\usage{
getSampleID(processed_counts, plate_list)
}
\arguments{
\item{processed_counts}{Output from `processCounts()` (reactive).}

\item{plate_list}{Plate name inside of the plate layout file.}
}
\value{
Returns the corresponding Sample ID for the correct row/column in
the plate layout file. Henceforth "Sample ID" refers to the code in the
plate layout file, while "Sample" is the code in the Luminex file.
}
\description{
A helper function to extract Sample ID based on plate name and row/col
}
\author{
Dionne Argyropoulos
}
