% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCard.R
\name{makeCard}
\alias{makeCard}
\title{Make Card in Fluent UI}
\usage{
makeCard(title, id, content, size = 12, style = "")
}
\arguments{
\item{title}{String with the large title that will be printed in the card.}

\item{id}{Identifying tag for use to link.}

\item{content}{A list of content to be rendered.}

\item{size}{A value from 1 to 12 of the width of the screen (default = 12).}

\item{style}{Value for any css styling (reactive).}
}
\value{
A "card" in the Fluent UI format with content.
}
\description{
This function imports the makes a card following the Fluent UI format.
}
