% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFItoRAU_LDH.R
\name{MFItoRAU_LDH}
\alias{MFItoRAU_LDH}
\title{Median Fluorescent Intensity (MFI) to Relative Antibody Units (RAU)
conversion for LDH}
\usage{
MFItoRAU_LDH(
  antigen_output,
  plate_list,
  dilution = c(1e+06, 333333.33, 111111.11, 37037.04, 12345.68, 4115.23, 1371.74, 457.25,
    152.42, 50.81),
  file_path = NULL
)
}
\arguments{
\item{antigen_output}{Output from `readAntigens()` or `readSeroData()` (reactive).}

\item{plate_list}{Output from `readPlateLayout()` (reactive).}

\item{dilution}{A list of numbers ranging from S1 to S10. Default: 1000000, 333333.33, 111111.11, 37037.04, 12345.68, 4115.23, 1371.74, 457.25, 152.42, 50.81.}

\item{file_path}{A file path to write the .csv final file. Default: Current working directory.}
}
\value{
A data frame containing the MFI and RAU Dilution values for each sample
}
\description{
This function fits a 5-parameter logistic standard curve to the dilutions
of the positive controls for each protein and converts the MFI values
into relative antibody units (RAU).
}
\author{
Connie Li Wai Suen, Caitlin Bourke, Dionne Argyropoulos
}
