% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFItoRAU_Plasmo.R
\name{MFItoRAU_PfPv}
\alias{MFItoRAU_PfPv}
\title{Median Fluorescent Intensity (MFI) to Relative Antibody Units (RAU)
conversion for either PNG or ETH standard}
\usage{
MFItoRAU_PfPv(
  processed_PfPv,
  plate_list,
  std_point,
  location,
  counts_QC_output
)
}
\arguments{
\item{processed_PfPv}{df$PfPv of output `processPkPfPv()`}

\item{plate_list}{Output of `readPlateLayout()`}

\item{std_point}{Standard Point Curve: 5 = 5-point curve, 10 = 10-point curve. Value is an integer.}

\item{location}{"PNG" or "ETH" to filter WEHI standard curve data (reactive).}

\item{counts_QC_output}{Output from `getCountsQC()`}
}
\value{
A list of three data frames:
1. Data frame with  MFI data, converted RAU data and matched SampleID's.
2. Plot information for `plotModel` function.
3. Data frame of RAU data for random forest classification use.
}
\description{
This function fits a 5-parameter logistic standard curve to the dilutions
of the positive controls for each protein and converts the MFI values
into relative antibody units (RAU). This function is utilised in the master
function `MFItoRAU_Plasmo()`.
}
\author{
Dionne Argyropoulos, Connie Li Wai Suen, Eamon Conway
}
