% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFItoRAU_Plasmo.R
\name{MFItoRAU_Pk}
\alias{MFItoRAU_Pk}
\title{Median Fluorescent Intensity (MFI) to Relative Antibody Units (RAU)
conversion for Pk proteins}
\usage{
MFItoRAU_Pk(processed_Pk, plate_list, std_point, counts_QC_output)
}
\arguments{
\item{processed_Pk}{df$Pk of output `processPkPfPv()`}

\item{plate_list}{Output of `readPlateLayout()`}

\item{std_point}{Standard Point Curve: 5 = 5-point curve, 10 = 10-point curve. Value is an integer.}

\item{counts_QC_output}{Output from `getCountsQC()`}
}
\value{
Data frame with MFI data, converted RAU data and matched SampleID's.
}
\description{
This function is utilised in the master function `MFItoRAU_Plasmo()`.
}
\author{
Dionne Argyropoulos, Caitlin Bourke
}
