% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFItoRAU.R
\name{MFItoRAU_Plasmo}
\alias{MFItoRAU_Plasmo}
\title{Median Fluorescent Intensity (MFI) to Relative Antibody Units (RAU)
conversion for Pk/Pf/Pv Master Function}
\usage{
MFItoRAU_Plasmo(sero_data, plate_list, panel = "panel1", std_point, qc_results)
}
\arguments{
\item{sero_data}{Output of `readserodata_output()`}

\item{plate_list}{Output of `readPlateLayout()`}

\item{panel}{Panel of Pk/Pf/Pv antigens. Default = "panel1" or user provided csv of Antigens and Species.}

\item{std_point}{Standard Point Curve: 5 = 5-point curve, 10 = 10-point curve, "PvLDH" for LDH specific curve. Default = 10. Value is an integer.}

\item{qc_results}{Output from `runQC()`.}
}
\value{
A list of three data frames:
1. Data frame with MFI data, converted RAU data, matched SampleID's, all intermediate dilution conversion factors
2. Data frame with only SampleID's, MFI and RAU data
3. Data frame #2 in long-format
}
\description{
This function leverages `MFItoRAU_Pk()` and `MFItoRAU()` to create a final MFI to RAU
output for Pk/Pf/Pv analyses.
}
\examples{
\donttest{
# Example demonstrating multi-plate 5-standard processing workflow.
# These files are included in the SeroTrackR package under inst/extdata.

your_raw_data_5std <- c(
  system.file("extdata", "example_MAGPIX_pk_5std_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_pk_5std_plate2.csv", package = "SeroTrackR")
)
your_plate_layout_5std <- system.file(
  "extdata", "example_platelayout_pk_5std.xlsx",
  package = "SeroTrackR"
)

# Read in raw MAGPIX data
sero_data <- readSeroData(
  raw_data = your_raw_data_5std,
  platform = "magpix"
)

# Read matching plate layout
plate_list <- readPlateLayout(
  plate_layout = your_plate_layout_5std,
  sero_data = sero_data
)
# Quality control
qc_results  <- runQC(sero_data, plate_list)

# Run MFI to RAU conversion
mfi_outputs               <- MFItoRAU_Plasmo(
  sero_data = sero_data,
  plate_list = plate_list,
  panel = "panel1",
  std_point = 5,
  qc_results = qc_results
)

# View All Outputs
mfi_outputs
}
}
\author{
Dionne Argyropoulos, Caitlin Bourke
}
