% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFItoRAU_Plasmo.R
\name{MFItoRAU_Plasmo}
\alias{MFItoRAU_Plasmo}
\title{Median Fluorescent Intensity (MFI) to Relative Antibody Units (RAU)
conversion for Pk/Pf/Pv Master Function}
\usage{
MFItoRAU_Plasmo(
  serodata_output,
  plate_list,
  panel = "panel1",
  std_point,
  counts_QC_output
)
}
\arguments{
\item{serodata_output}{Output of `readserodata_output()`}

\item{plate_list}{Output of `readPlateLayout()`}

\item{panel}{Panel of Pk/Pf/Pv antigens. Default = "panel1".}

\item{std_point}{Standard Point Curve: 5 = 5-point curve, 10 = 10-point curve. Value is an integer.}

\item{counts_QC_output}{Output from `getCountsQC()`.}
}
\value{
A list of three data frames:
1. Data frame with MFI data, converted RAU data, matched SampleID's, all intermediate dilution conversion factors
2. Data frame with only SampleID's, MFI and RAU data
3. Data frame #2 in long-format
}
\description{
This function leverages `MFItoRAU_Pk()` and `MFItoRAU_PfPv()` to create a final MFI to RAU
output for Pk/Pf/Pv analyses.
}
\author{
Dionne Argyropoulos
}
