% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBeadCounts.R
\name{plotBeadCounts}
\alias{plotBeadCounts}
\title{Plot Bead Counts per Plate per Antigen}
\usage{
plotBeadCounts(antigen_counts_output)
}
\arguments{
\item{antigen_counts_output}{Output from `getAntigenCounts()` (reactive).}
}
\value{
Dot plot with values > 15 threshold coloured in blue (sufficient
beads) and less than or equal to 15 beads coloured in red (repeat)
faceted by each antigen
(ggplot).
}
\description{
Enhances the `plotCounts()` output by providing greater resolution,
displaying antigens per plate, and enabling SampleID name visibility via
hover (transformed to Plotly in server.R)
}
\author{
Dionne Argyropoulos
}
