% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts.R
\name{plotCounts}
\alias{plotCounts}
\title{Plot Bead Count Data}
\usage{
plotCounts(qc_results, experiment_name)
}
\arguments{
\item{qc_results}{Output from `runQC()`.}

\item{experiment_name}{User-input experiment name.}
}
\value{
Tile Plot showing binary result of "sufficient beads" with cut-off
>15 beads and "repeat" less than or equal to 15 beads (ggplot).
}
\description{
This function gets the count data and plots the plate image, creating a new
facet (i.e., panel) for each antigen and each line represents the
different plates so that they can be visualised.
}
\examples{
\donttest{

# Step 0: Load example raw data
your_raw_data <- c(
  system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate2.csv", package = "SeroTrackR")
)
your_plate_layout <- system.file(
  "extdata",
  "example_platelayout_1.xlsx",
  package = "SeroTrackR"
)

# Step 1: Read serology data and plate layout
sero_data  <- readSeroData(your_raw_data,"magpix")
plate_list <- readPlateLayout(your_plate_layout, sero_data)

# Step 2: Process counts and perform quality control
qc_results  <- runQC(sero_data, plate_list)

# Step 3: Plot Counts
plotCounts(qc_results, "experiment1")
}
}
\author{
Shazia Ruybal-Pesántez, Dionne Argyropoulos
}
