% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMFI.R
\name{plotMFI}
\alias{plotMFI}
\title{Median Fluorescent Intensity (MFI) Box Plots}
\usage{
plotMFI(mfi_to_rau_output, location)
}
\arguments{
\item{mfi_to_rau_output}{Output from `MFItoRAU()` or `MFItoRAU_Adj()`
.}

\item{location}{"PNG" or "ETH".}
}
\value{
Box plots with MFI values for each protein (ggplot).
}
\description{
Boxplot of the MFI values.
}
\examples{
\donttest{

# Step 0: Load example raw data
your_raw_data <- c(
  system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate2.csv", package = "SeroTrackR")
)
your_plate_layout <- system.file(
  "extdata",
  "example_platelayout_1.xlsx",
  package = "SeroTrackR"
)

# Step 1: Read serology data and plate layout
sero_data  <- readSeroData(your_raw_data,"magpix")
plate_list <- readPlateLayout(your_plate_layout, sero_data)

# Step 2: Process counts and perform quality control
qc_results <- runQC(sero_data, plate_list)

# Step 3: Convert MFI to RAU using ETH beads
mfi_to_rau <- MFItoRAU_Adj(
  sero_data   = sero_data,
  plate_list  = plate_list,
  qc_results  = qc_results
)

# Step 4: Plot MFI values
plotMFI(mfi_to_rau, "MFI")
}
}
\author{
Dionne Argyropoulos
}
