% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMFI.R
\name{plotMFI}
\alias{plotMFI}
\title{Median Fluorescent Intensity (MFI) Box Plots}
\usage{
plotMFI(mfi_to_rau_output, location)
}
\arguments{
\item{mfi_to_rau_output}{Output from `MFItoRAU_PNG()` or `MFItoRAU_ETH()`
(reactive).}

\item{location}{"PNG" or "ETH" (reactive).}
}
\value{
Box plots with MFI values for each protein (ggplot).
}
\description{
Boxplot of the MFI values.
}
\author{
Dionne Argyropoulos
}
