% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModel_ETH.R
\name{plotModel_ETH}
\alias{plotModel_ETH}
\title{Plot the Median Fluorescent Intensity (MFI) to Relative Antibody Units (RAU)
Results Data based on ETH standard}
\usage{
plotModel_ETH(mfi_to_rau_output, antigen_output)
}
\arguments{
\item{mfi_to_rau_output}{Output from `MFItoRAU_ETH()` (reactive).}

\item{antigen_output}{Output from `readAntigens()` (reactive).}
}
\value{
List of dot and line plots of MFI to RAU model standard curve,
with each one representing an individual plate (ggplot).
}
\description{
This function gets the Median Fluorescent Intensity (MFI) to Relative
Antibody Units (RAU) model results data and plots the model fits based on
`MFItoRAU_ETH.`
}
\author{
Dionne Argyropoulos
}
