% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModel_PNG.R
\name{plotModel_PNG}
\alias{plotModel_PNG}
\title{Plot the Median Fluorescent Intensity (MFI) to Relative Antibody Units (RAU)
Results Data based on PNG standard}
\usage{
plotModel_PNG(mfi_to_rau_output, antigen_output)
}
\arguments{
\item{mfi_to_rau_output}{Output from `MFItoRAU_PNG()` (reactive).}

\item{antigen_output}{Output from `readAntigens()` (reactive).}
}
\value{
List of dot and line plots of MFI to RAU model standard curve,
with each one representing an individual plate (ggplot).
}
\description{
This function gets the Median Fluorescent Intensity (MFI) to Relative
Antibody Units (RAU) model results data and plots the model fits based on
`MFItoRAU_PNG.`
}
\author{
Shazia Ruybal-Pesantez, Dionne Argyropoulos
}
