% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRAU.R
\name{plotRAU}
\alias{plotRAU}
\title{Relative Antibody Unit (RAU) Box Plots}
\usage{
plotRAU(mfi_to_rau_output, location)
}
\arguments{
\item{mfi_to_rau_output}{Output from `MFItoRAU_PNG()` or `MFItoRAU_ETH()`
(reactive).}

\item{location}{"PNG" or "ETH" (reactive).}
}
\value{
Box plots with RAU values for each protein (ggplot).
}
\description{
Boxplot of the RAU values.
}
\author{
Dionne Argyropoulos
}
