% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStds.R
\name{plotStds}
\alias{plotStds}
\title{Plot Raw Median Fluorescent Intensity of Standard Curve Data}
\usage{
plotStds(antigen_output, location, experiment_name)
}
\arguments{
\item{antigen_output}{Output from `readAntigens()` (reactive).}

\item{location}{"PNG" or "ETH" to filter WEHI standard curve data (reactive).}

\item{experiment_name}{User-input experiment name (reactive).}
}
\value{
- Dot and line plot of standard curves (S1-S10) with PNG or Ethiopia stds
underneath (ggplot).
- WEHI-acceptable standard curve data on background of plot with user data.
}
\description{
This function gets the standards data and plots the standard curves.
}
\author{
Shazia Ruybal-Pesantez, Dionne Argyropoulos
}
