% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStds_PkPfPv.R
\name{plotStds_PkPfPv}
\alias{plotStds_PkPfPv}
\title{Plot Raw Median Fluorescent Intensity of Pk/Pf/Pv Standard Curve Data}
\usage{
plotStds_PkPfPv(serodata_output, experiment_name)
}
\arguments{
\item{serodata_output}{Output from `readSeroData()` (reactive).}

\item{experiment_name}{User-input experiment name (reactive).}
}
\value{
- Dot and line plot of standard curves (S1-S10)
- WEHI-acceptable standard curve data on background of plot with user data.
}
\description{
This function gets the standards data and plots the standard curves for antigens in the Pk/Pf/Pv panel.
}
\author{
Dionne Argyropoulos
}
