% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStds.R
\name{plotStds_PkPfPv}
\alias{plotStds_PkPfPv}
\title{Plot Raw Median Fluorescent Intensity of Pk/Pf/Pv Standard Curve Data}
\usage{
plotStds_PkPfPv(sero_data, experiment_name, panel)
}
\arguments{
\item{sero_data}{Output from `readSeroData()`.}

\item{experiment_name}{User-input experiment name.}

\item{panel}{Panel of Pk/Pf/Pv antigens. Default = "panel1" or user provided csv of Antigens and Species.}
}
\value{
- Dot and line plot of standard curves (S1-S10)
- WEHI-acceptable standard curve data on background of plot with user data.
}
\description{
This function gets the standards data and plots the standard curves for antigens in the Pk/Pf/Pv panel.
}
\examples{
\donttest{
# Example demonstrating how to process bead count data.
# These files are included in the SeroTrackR package under inst/extdata.

your_raw_data <- c(
   system.file("extdata", "example_MAGPIX_pk_5std_plate1.csv", package = "SeroTrackR"),
   system.file("extdata", "example_MAGPIX_pk_5std_plate2.csv", package = "SeroTrackR")
)

# Read in raw MAGPIX data
sero_data <- readSeroData(
  raw_data = your_raw_data,
  platform = "magpix"
)

# Plot Standards
plotStds_PkPfPv(
  sero_data = sero_data,
  experiment_name = "experiment1"
)

}

}
\author{
Dionne Argyropoulos
}
