% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQC.R
\name{processCounts}
\alias{processCounts}
\title{Process Counts from Raw Serological Data file}
\usage{
processCounts(sero_data)
}
\arguments{
\item{sero_data}{Output from `readSeroData()`.}
}
\value{
Returns a long table of counts with "Warning" category (<15 == 1 and
\eqn{>=}  15 == 0) for downstream wrangling.
}
\description{
A helper function to process counts data.
}
\examples{
\donttest{
# Example demonstrating how to process bead count data.
# These files are included in the SeroTrackR package under inst/extdata.

your_raw_data <- c(
  system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate2.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate3.csv", package = "SeroTrackR")
)

# Read in raw MAGPIX data
sero_data <- readSeroData(
  raw_data = your_raw_data,
  platform = "magpix"
)

# Process counts
processed_master <- processCounts(sero_data = sero_data)

}

}
\author{
Dionne Argyropoulos
}
