% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCounts.R
\name{processCounts}
\alias{processCounts}
\title{Process Counts from Raw Serological Data file}
\usage{
processCounts(antigen_output)
}
\arguments{
\item{antigen_output}{Output from `readAntigens()` (reactive).}
}
\value{
Returns a long table of counts with "Warning" category (<15 == 1 and
≥ 15 == 0) for downstream wrangling.
}
\description{
A helper function to process counts data.
}
\author{
Dionne Argyropoulos
}
