% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processPkPfPv.R
\name{processPkPfPv}
\alias{processPkPfPv}
\title{Processing Serological Data for Pk/Pf/Pv MFI to RAU conversion}
\usage{
processPkPfPv(sero_data, plate_list, panel = "panel1")
}
\arguments{
\item{sero_data}{Output of `readSeroData()`}

\item{plate_list}{Output of `readPlateLayout()`}

\item{panel}{Panel of Pk/Pf/Pv antigens. Default = "panel1" or user provided csv of Antigens and Species.}
}
\value{
A list of two data frames:
1. Data frame with Pk antigens
2. Data frame with Pf/Pv antigens
}
\description{
This is a pre-requisite function before running the `MFItoRAU_Plasmo()` so that the
appropriate MFI to RAU conversions can be run for the respective antigens.
}
\examples{
\donttest{
# Example demonstrating multi-plate 5-standard processing workflow.
# These files are included in the SeroTrackR package under inst/extdata.

your_raw_data_5std <- c(
  system.file("extdata", "example_MAGPIX_pk_5std_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_pk_5std_plate2.csv", package = "SeroTrackR")
)

your_plate_layout_5std <- system.file(
  "extdata", "example_platelayout_pk_5std.xlsx",
  package = "SeroTrackR"
)

# Read in raw MAGPIX data
sero_data <- readSeroData(
  raw_data = your_raw_data_5std,
  platform = "magpix"
)

# Read matching plate layout
plate_list <- readPlateLayout(
  plate_layout = your_plate_layout_5std,
  sero_data = sero_data
)

# Process multi-species panel
processed_master <- processPkPfPv(
  sero_data = sero_data,
  plate_list = plate_list,
  panel = "panel1"
)
}
}
\author{
Dionne Argyropoulos
}
