% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvsero-package.R
\docType{package}
\name{pvsero-package}
\alias{pvsero}
\alias{pvsero-package}
\title{pvsero: Functions for the Sero Analysis}
\description{
Data wrangling and cleaning, quality control checks and implementation of machine learning classification algorithm.
}
\author{
\strong{Maintainer}: Dionne Argyropoulos \email{argyropoulos.d@wehi.edu.au}

}
\keyword{internal}
