% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeroData.R
\name{readSeroData}
\alias{readSeroData}
\title{Read Raw Serological Data}
\usage{
readSeroData(raw_data, platform, version = "4.2", raw_data_filenames = NULL)
}
\arguments{
\item{raw_data}{String with the raw data path (reactive).}

\item{platform}{"magpix", "bioplex" or "intelliflex" (reactive).}

\item{version}{xPONENT software version. For "magpix" can be 4.2 or 4.3 (reactive). Default: 4.2.}

\item{raw_data_filenames}{String with the raw data filename path (reactive).
Default is NA as it can be deduced from raw_data. Needs to be a parameter for the PvSeroApp.}
}
\value{
List of data frames: (i) raw data output, (ii) cleaned all results
(iii) count data, (iv) blanks only, (v) standards only, (vi) run
information.
}
\description{
This function imports the raw data from the Magpix or Bioplex machine
and matches the sample names from the plate layout based on their plate/well
location.
}
\examples{

# Example raw data files (MAGPIX platform)
your_raw_data <- c(
  system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate2.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate3.csv", package = "SeroTrackR")
)

# Read and combine raw serology data
sero_data <- readSeroData(
  raw_data = your_raw_data,
  platform = "magpix"
)
}
\author{
Dionne Argyropoulos, Shazia Ruybal-Pesántez
}
