% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeroData.R
\name{readSeroData}
\alias{readSeroData}
\title{Read Raw Serological Data}
\usage{
readSeroData(raw_data, platform)
}
\arguments{
\item{raw_data}{String with the raw data path (reactive).}

\item{platform}{"magpix", "bioplex" or "intelliflex" (reactive).}
}
\value{
List of data frames: (i) raw data output, (ii) cleaned all results
(iii) count data, (iv) blanks only, (v) standards only, (vi) run
information.
}
\description{
This function imports the raw data from the Magpix or Bioplex machine
and matches the sample names from the plate layout based on their plate/well
location.
}
\author{
Shazia Ruybal-Pesántez, Dionne Argyropoulos
}
