% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderClassificationTable.R
\name{renderClassificationTable}
\alias{renderClassificationTable}
\title{All Classification Data}
\usage{
renderClassificationTable(mfi_to_rau_output, algorithm_type, qc_results)
}
\arguments{
\item{mfi_to_rau_output}{Output from `MFItoRAU()` or `MFItoRAU_Adj()`.}

\item{algorithm_type}{Algorithm: "antibody_model" (PvSEM algorithm; default)}

\item{qc_results}{Output from `runQC()`.}
}
\value{
A table of all classification outputs.
}
\description{
This function runs the classification algorithm for all possible sensitivity
and specificity options.
}
\examples{
\donttest{

# Step 0: Load in Raw Data
your_raw_data <- c(
  system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate2.csv", package = "SeroTrackR")
)
your_plate_layout <- system.file("extdata", "example_platelayout_1.xlsx", package = "SeroTrackR")

# Step 1: Reading in Raw Data
sero_data     <- readSeroData(raw_data = your_raw_data, "magpix")
plate_list    <- readPlateLayout(
  plate_layout = your_plate_layout,
  sero_data = sero_data
)

# Step 2: Quality Control and MFI to RAU
qc_results <- runQC(sero_data, plate_list)

# Step 4: Run MFI to RAU (e.g., using ETH beads)
mfi_to_rau_output  <- MFItoRAU_Adj(sero_data, plate_list, qc_results)

# Step 5: Render classification table
renderClassificationTable(
  mfi_to_rau_output = mfi_to_rau_output,
  algorithm_type = "antibody_model",
  qc_results = qc_results
)

}

}
\author{
Dionne Argyropoulos
}
