% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderClassificationTable.R
\name{renderClassificationTable}
\alias{renderClassificationTable}
\title{All Classification Data}
\usage{
renderClassificationTable(mfi_to_rau_output, algorithm_type, counts_QC_output)
}
\arguments{
\item{mfi_to_rau_output}{Output from `MFItoRAU_PNG()` or `MFItoRAU_ETH()`
(reactive).}

\item{algorithm_type}{User-selected algorithm choice:
- "antibody_model" (PvSeroTaT model; default), or
- "antibody_model_excLF016" (PvSeroTat excluding LF016).}

\item{counts_QC_output}{Output from `getCountsQC()` (reactive).}
}
\value{
A table of all classification outputs.
}
\description{
This function runs the classification algorithm for all possible sensitivity
and specificity options.
}
\author{
Dionne Argyropoulos
}
