% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderDetailsList.R
\name{renderDetailsList}
\alias{renderDetailsList}
\title{Create a Fluent UI Table}
\usage{
renderDetailsList(df)
}
\arguments{
\item{df}{Any processed data frame}
}
\value{
A table in the Fluent UI format
}
\description{
This function makes the table in a Fluent UI format.
}
\author{
Dionne Argyropoulos
}
