% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderQCReport.R
\name{renderQCReport}
\alias{renderQCReport}
\title{Generate QC PDF Report}
\usage{
renderQCReport(
  raw_data,
  plate_layout,
  platform,
  experiment_name = "experiment1",
  date = format(Sys.Date(), "\%Y\%m\%d"),
  experiment_notes = "no notes",
  location,
  std_point = 10,
  path = "."
)
}
\arguments{
\item{raw_data}{A string with the raw data path.}

\item{plate_layout}{A string with the plate layout path.}

\item{platform}{A string: "magpix", "intelliflex", or "bioplex".}

\item{experiment_name}{A string for experiment name.}

\item{date}{A string or Date. Defaults to today's date.}

\item{experiment_notes}{A string of notes. Default is "no notes".}

\item{location}{A string for experiment location: "ETH" or "PNG" accepted.}

\item{std_point}{Standard Point Curve: 5 = 5-point curve, 10 = 10-point curve, "PvLDH" for LDH specific curve. Default = 10. Value is an integer.}

\item{path}{Output path for the PDF file. Defaults to current working directory.}
}
\value{
Rendered PDF report.
}
\description{
Generate QC PDF Report
}
\examples{
## Not run on CRAN because it requires interactive rendering and can be slow:
\dontrun{
  # Example raw data files (MAGPIX platform)
  your_raw_data <- c(
    system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR"),
    system.file("extdata", "example_MAGPIX_plate2.csv", package = "SeroTrackR"),
    system.file("extdata", "example_MAGPIX_plate3.csv", package = "SeroTrackR")
  )

  # Example plate layout file
  your_plate_layout <- system.file(
    "extdata",
    "example_platelayout_1.xlsx",
    package = "SeroTrackR"
  )

  # Generate the QC PDF report
  renderQCReport(
    raw_data     = your_raw_data,
    plate_layout = your_plate_layout,
    platform     = "magpix",
    location     = "ETH"
  )
}
}
\author{
Dionne Argyropoulos
}
