% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderQCReport.R
\name{renderQCReport}
\alias{renderQCReport}
\title{Generate QC PDF Report}
\usage{
renderQCReport(
  raw_data,
  plate_layout,
  platform,
  experiment_name = "experiment1",
  date = format(Sys.Date(), "\%Y\%m\%d"),
  experiment_notes = "no notes",
  location,
  path = "."
)
}
\arguments{
\item{raw_data}{A string with the raw data path.}

\item{plate_layout}{A string with the plate layout path.}

\item{platform}{A string: "magpix", "intelliflex", or "bioplex".}

\item{experiment_name}{A string for experiment name.}

\item{date}{A string or Date. Defaults to today's date.}

\item{experiment_notes}{A string of notes. Default is "no notes".}

\item{location}{A string for experiment location: "ETH" or "PNG" accepted.}

\item{path}{Output path for the PDF file. Defaults to current working directory.}
}
\value{
Rendered PDF report.
}
\description{
Generate QC PDF Report
}
\author{
Dionne Argyropoulos
}
