% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderReport.R
\name{renderReport}
\alias{renderReport}
\title{Render Markdown report}
\usage{
renderReport(input, output, params)
}
\arguments{
\item{input}{Input files}

\item{output}{Output files}

\item{params}{Parameters to generate outputs.}
}
\value{
PDF output.
}
\description{
A short function to render the rmarkdown report on Shiny.
}
\author{
Dionne Argyropoulos
}
