% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderTwoCols.R
\name{renderTwoCols}
\alias{renderTwoCols}
\title{Create two columns in Fluent UI}
\usage{
renderTwoCols(
  first_col,
  second_col,
  first_width = "50\%",
  second_width = "50\%"
)
}
\arguments{
\item{first_col}{A list of content for the first column.}

\item{second_col}{A list of content for the second column.}

\item{first_width}{Percent width of the column space (default: 50\%).}

\item{second_width}{Percent width of the column space (default: 50\%).}
}
\value{
Fluent UI window with two columns.
}
\description{
This function creates two columns in the Fluent UI format.
}
