% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPlasmoPipeline.R
\name{runPlasmoPipeline}
\alias{runPlasmoPipeline}
\title{Run Pk/Pf/Pv Data Analysis Pipeline from Start to End}
\usage{
runPlasmoPipeline(
  raw_data,
  platform = "magpix",
  plate_layout,
  panel = "panel1",
  std_point,
  experiment_name = "experiment1",
  algorithm_type = "antibody_model",
  sens_spec = "maximised"
)
}
\arguments{
\item{raw_data}{String with the raw data path.}

\item{platform}{"magpix" or "bioplex". Default: "Bioplex"}

\item{plate_layout}{An ".xlsx" file with sheets labelled plate1, plate2... etc.}

\item{panel}{Panel of Pk/Pf/Pv antigens. Default = "panel1".}

\item{std_point}{Standard Point Curve: 5 = 5-point curve, 10 = 10-point curve. Value is an integer.}

\item{experiment_name}{User-input experiment name. Default: "experiment1".}

\item{algorithm_type}{User-selected algorithm choice:
- "antibody_model" (PvSeroTaT model; default), or
- "antibody_model_excLF016" (PvSeroTaT excluding LF016).}

\item{sens_spec}{User-selected Sensitivity/Specificity threshold:
- "maximised" (default),
- "85\% sensitivity",
- "90\% sensitivity",
- "95\% sensitivity",
- "85\% specificity",
- "90\% specificity".
- "95\% specificity".}
}
\value{
A data frame containing the MFI and RAU Dilution values for each sample, QC plots for standard curve, bead counts and blanks.
}
\description{
Run Pk/Pf/Pv Data Analysis Pipeline from Start to End
}
\author{
Dionne Argyropoulos
}
