% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPlasmoPipeline.R
\name{runPlasmoPipeline}
\alias{runPlasmoPipeline}
\title{Run Pk/Pf/Pv Data Analysis Pipeline from Start to End}
\usage{
runPlasmoPipeline(
  raw_data,
  platform = "magpix",
  plate_layout,
  panel = "panel1",
  std_point,
  experiment_name = "experiment1",
  algorithm_type = "antibody_model",
  sens_spec = "maximised"
)
}
\arguments{
\item{raw_data}{String with the raw data path.}

\item{platform}{"magpix" or "bioplex". Default: "Bioplex"}

\item{plate_layout}{An ".xlsx" file with sheets labelled plate1, plate2... etc.}

\item{panel}{Panel of Pk/Pf/Pv antigens. Default = "panel1".}

\item{std_point}{Standard Point Curve: 5 = 5-point curve, 10 = 10-point curve. Value is an integer.}

\item{experiment_name}{User-input experiment name. Default: "experiment1".}

\item{algorithm_type}{User-selected algorithm choice:
- "antibody_model" (PvSeroTaT model; default), or
- "antibody_model_excLF016" (PvSeroTaT excluding LF016).}

\item{sens_spec}{User-selected Sensitivity/Specificity threshold:
- "maximised" (default),
- "85\% sensitivity",
- "90\% sensitivity",
- "95\% sensitivity",
- "85\% specificity",
- "90\% specificity".
- "95\% specificity".}
}
\value{
A data frame containing the MFI and RAU Dilution values for each sample, QC plots for standard curve, bead counts and blanks.
}
\description{
Run Pk/Pf/Pv Data Analysis Pipeline from Start to End
}
\examples{
\donttest{

# Helper to avoid repetition in examples
run_example_std <- function(std_point) {
  # Load raw data for given standard curve
  your_raw_data <- c(
    system.file("extdata",
                paste0("example_MAGPIX_pk_", std_point, "std_plate1.csv"),
                package = "SeroTrackR"),
    system.file("extdata",
                paste0("example_MAGPIX_pk_", std_point, "std_plate2.csv"),
                package = "SeroTrackR")
  )

  layout_file <- system.file(
    "extdata",
    paste0("example_platelayout_pk_", std_point, "std.xlsx"),
    package = "SeroTrackR"
  )

  # Run pipeline
  runPlasmoPipeline(
    raw_data = your_raw_data,
    platform = "magpix",
    plate_layout = layout_file,
    panel = "panel1",
    std_point = std_point,
    experiment_name = paste0(std_point, "-point standard curve")
  )
}

# ---- 5-point standard curve ----
results_5std <- run_example_std(5)

# ---- 10-point standard curve ----
results_10std <- run_example_std(10)

}
}
\author{
Dionne Argyropoulos
}
