% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPvSeroPipeline.R
\name{runPvSeroPipeline}
\alias{runPvSeroPipeline}
\title{Run PvSero Pipeline from Start to End}
\usage{
runPvSeroPipeline(
  raw_data,
  plate_layout,
  platform = "magpix",
  location,
  experiment_name = "experiment1",
  std_point = 10,
  classify = "Yes",
  algorithm_type = "antibody_model",
  sens_spec = "balanced"
)
}
\arguments{
\item{raw_data}{String with the raw data path.}

\item{plate_layout}{An ".xlsx" file with sheets labelled plate1, plate2... etc.}

\item{platform}{"magpix" or "bioplex". Default = "magpix".}

\item{location}{"PNG" or "ETH" to filter WEHI standard curve data.}

\item{experiment_name}{User-input experiment name.}

\item{std_point}{Standard Point Curve: 5 = 5-point curve, 10 = 10-point curve, "PvLDH" for LDH specific curve. Default = 10. Value is an integer.}

\item{classify}{"Yes" or "No" depending on whether you would like classification or not. Default = "Yes".}

\item{algorithm_type}{Algorithm: "antibody_model" (PvSEM algorithm; default)}

\item{sens_spec}{User-selected Sensitivity/Specificity threshold: "balanced"
(default) or "90\% specificity".}
}
\value{
classifyResults_output, stdcurve_plot, plateqc_plot, check_repeats_output, blanks_plot, model_plot
}
\description{
A master function combining the entire PvSeroApp pipeline into one command to run in R.
}
\examples{

\donttest{
# Example data supplied with the package
your_raw_data <- c(
  system.file("extdata", "example_MAGPIX_plate1.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate2.csv", package = "SeroTrackR"),
  system.file("extdata", "example_MAGPIX_plate3.csv", package = "SeroTrackR")
)

plate_layout <- system.file(
  "extdata", "example_platelayout_1.xlsx", package = "SeroTrackR"
)

# Run full pipeline including classification
runPvSeroPipeline(
  raw_data = your_raw_data,
  plate_layout = plate_layout,
  platform = "magpix",
  location = "PNG",
  experiment_name = "experiment1",
  std_point = 10,
  algorithm_type = "antibody_model",
  sens_spec = "balanced",
  classify = "Yes"
)

# Run processing pipeline only (no classification)
runPvSeroPipeline(
  raw_data = your_raw_data,
  plate_layout = plate_layout,
  platform = "magpix",
  location = "PNG",
  experiment_name = "experiment1",
  std_point = 10,
  algorithm_type = "antibody_model",
  sens_spec = "balanced",
  classify = "No"
)
}
}
\author{
Dionne Argyropoulos
}
