% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPvSeroPipeline.R
\name{runPvSeroPipeline}
\alias{runPvSeroPipeline}
\title{Run PvSero Pipeline from Start to End}
\usage{
runPvSeroPipeline(
  raw_data,
  raw_data_filenames,
  platform,
  location,
  experiment_name,
  algorithm_type,
  sens_spec
)
}
\arguments{
\item{raw_data}{String with the raw data path (reactive).}

\item{raw_data_filenames}{String with the raw data filenames (reactive).}

\item{platform}{"magpix" or "bioplex" (reactive).}

\item{location}{"PNG" or "ETH" to filter WEHI standard curve data (reactive).}

\item{experiment_name}{User-input experiment name (reactive).}

\item{algorithm_type}{User-selected algorithm choice:
- "antibody_model" (PvSeroTaT model; default), or
- "antibody_model_excLF016" (PvSeroTat excluding LF016).}

\item{sens_spec}{User-selected Sensitivity/Specificity threshold:
- "maximised" (default),
- "85\% sensitivity",
- "90\% sensitivity",
- "95\% sensitivity",
- "85\% specificity",
- "90\% specificity".
- "95\% specificity".}
}
\value{
classifyResults_output, stdcurve_plot, plateqc_plot, check_repeats_output, blanks_plot, model_plot
}
\description{
A master function combining the entire PvSeroApp pipeline into one command to run in R.
}
\author{
Dionne Argyropoulos
}
