% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeroTrackR-package.R
\docType{package}
\name{SeroTrackR-package}
\alias{SeroTrackR}
\alias{SeroTrackR-package}
\title{SeroTrackR: Serology-Based Data Analysis and Visualization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Data wrangling and cleaning, quality control checks and implementation of machine learning classification algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dionnecargy/SeroTrackR}
  \item \url{https://dionnecargy.github.io/SeroTrackR/}
  \item Report bugs at \url{https://github.com/dionnecargy/SeroTrackR/issues}
}

}
\author{
\strong{Maintainer}: Dionne Argyropoulos \email{argyropoulos.d@wehi.edu.au}

Authors:
\itemize{
  \item Dionne Argyropoulos \email{argyropoulos.d@wehi.edu.au}
}

}
\keyword{internal}
