% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serotrackr-package.R
\docType{package}
\name{serotrackr-package}
\alias{serotrackr}
\alias{serotrackr-package}
\title{serotrackr: Facilitates Submissions to SeroTracker}
\description{
serotrackr is a package specifically developed to help users prepare their raw data for submission to the SeroTracker platform. This happens through several steps represented by one function each; 1) mapping the columns from raw data to those required for submission, 2) cleaning the raw data, 3) generating aggregate estimayes, 4) adjusting the estimates, 5) exporting the final data to a xlsx file where the remaining fields can be filled out by the user before submission on SeroTracker.com.
}
\keyword{internal}
