% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serotrackr-package.R
\docType{package}
\name{serotrackr-package}
\alias{serotrackr}
\alias{serotrackr-package}
\title{serotrackr: Help with SeroTracker Submissions}
\description{
serotrackr is an R package that helps users prepare their raw data for submission to the SeroTracker platform. It has functions to 1) validate the raw data, 2) locate region names and convert them to standard region codes, 3) aggregate data into estimates, and 4) save the results to an Excel document, which will then be submitted on SeroTracker.com.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://serotracker.github.io/serotrackr}
  \item \url{https://github.com/serotracker/serotrackr}
  \item Report bugs at \url{https://github.com/serotracker/serotrackr/issues}
}

}
\author{
\strong{Maintainer}: Reza Hosseini \email{rezahs@alum.ubc.ca} (\href{https://orcid.org/0000-0002-2390-4884}{ORCID})

Other contributors:
\itemize{
  \item SeroTracker [copyright holder, funder]
}

}
\keyword{internal}
