% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serotrackr-package.R
\docType{package}
\name{serotrackr-package}
\alias{serotrackr}
\alias{serotrackr-package}
\title{serotrackr: Help with SeroTracker Submissions}
\description{
serotrackr is a package specifically developed to help users prepare their raw data for submission to the SeroTracker platform. This happens through several steps represented by one function each; 1) mapping the columns from raw data to those required for submission, 2) cleaning the raw data, 3) generating aggregate estimayes, 4) adjusting the estimates, 5) exporting the final data to a xlsx file where the remaining fields can be filled out by the user before submission on SeroTracker.com.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://serotracker.github.io/serotrackr}
  \item \url{https://github.com/serotracker/serotrackr}
  \item Report bugs at \url{https://github.com/serotracker/serotrackr/issues}
}

}
\author{
\strong{Maintainer}: Reza Hosseini \email{rezahs@alum.ubc.ca} (\href{https://orcid.org/0000-0002-2390-4884}{ORCID})

Other contributors:
\itemize{
  \item SeroTracker [copyright holder, funder]
}

}
\keyword{internal}
